Major restructure, breaking out functionality into subpackages
