+++ msm_we/msm_we.py
@@ -1,28 +1,29 @@
 """haMSM estimation and analysis"""
 from __future__ import division, print_function
-
 __metaclass__ = type
 import numpy as np
 import tqdm.auto as tqdm
 from functools import partialmethod
-import sys
-import h5py
 import concurrent
 import multiprocessing as mp
 from copy import deepcopy
-from westpa import analysis
-from westpa.core.binning import RectilinearBinMapper, VoronoiBinMapper
-
-from scipy.sparse import coo_matrix
-import scipy.sparse as sparse
-from sklearn.decomposition import IncrementalPCA as iPCA
-from sklearn.cluster import KMeans as kmeans
-from sklearn.cluster import MiniBatchKMeans as mini_kmeans
+import mdtraj as md
+from rich.live import Live
+from rich.table import Table
+import ray
 from sklearn.utils._openmp_helpers import _openmp_effective_n_threads
-from deeptime.decomposition import TICA, VAMP
+# My mixin structure may be a little strange here, because these aren't really mixins that are meant to be
+#   reused, it's just to break these out.
+# Alternatively, I think I could drop the functions straight in those modules without the mixin classes,
+#   and within my modelWE definition just do (for example)
+#   class modelWE:
+#       from ._clustering import *
 
-from msm_we.stratified_clustering import StratifiedClusters
-from msm_we.utils import is_connected, find_connected_sets, inverse_iteration
+from msm_we._clustering import ClusteringMixin
+from msm_we._dimensionality import DimensionalityReductionMixin
+from msm_we._plotting import PlottingMixin
+from msm_we._analysis import AnalysisMixin
+from msm_we._data import DataMixin
 
 import logging
 from rich.logging import RichHandler
@@ -32,24 +33,8 @@ log.setLevel(logging.INFO)
 log.addHandler(RichHandler())
 log.propagate = False
 
-# Using the tkinter backend makes matplotlib run better on a cluster, maybe?
-# import matplotlib
-# matplotlib.use("TkAgg")
-import matplotlib.pyplot as plt
 
-import mdtraj as md
-
-from rich.live import Live
-from rich.table import Table
-
-import ray
-
-# If you implement a custom bin mapper that should work with stratified clustering, you can add it to this set
-#   after importing msm_we
-SUPPORTED_MAPPERS = {RectilinearBinMapper, VoronoiBinMapper}
-
-
-class modelWE:
+class modelWE(ClusteringMixin, DimensionalityReductionMixin, PlottingMixin, AnalysisMixin, DataMixin):
     """
     History-augmented Markov state model estimation from WE data
 
@@ -188,7 +173,6 @@ class modelWE:
 
         self.errorWeight = None
         self.errorCount = None
-        self.fluxMatrixRaw = None
 
         self.targetRMSD_centers = None
         """array-like: List of RMSDs corresponding to each cluster."""
@@ -1057,239 +1041,6 @@ class modelWE:
         self.validation_iterations = validation_iterations
         self.validation_models = validation_models
 
-    def load_iter_data(self, n_iter: int):
-        """
-        Update state with the data (including pcoord but not including coords) corresponding to an iteration.
-
-        Object fields updated with the information from the selected iteration:
-            - `self.westList`
-            - `self.segindList`
-            - `self.weightList`
-            - `self.n_segs`
-            - `self.pcoord0List`
-            - `self.pcoord1List`
-
-        Parameters
-        ----------
-        n_iter : int
-            Iteration to get data for.
-
-        Returns
-        -------
-        None
-
-        Todo
-        ----
-        May want to rework the logic here, depending on how this is used.
-        Seems like some of this iteration can be removed/optimized.
-        """
-
-        # log.debug("Getting iteration data")
-
-        self.n_iter = n_iter
-        westList = np.array([])
-        segindList = np.array([])
-        weightList = np.array([])
-        pcoord0List = np.empty((0, self.pcoord_ndim))
-        pcoord1List = np.empty((0, self.pcoord_ndim))
-
-        seg_weights = np.array([])
-
-        n_segs = 0
-
-        # Iterate through each file index, trying to find files that contains the iteration of interest
-        # TODO: Can replace this with `for if, fileName in enumerate(self.\\)`
-        for file_idx in range(self.n_data_files):
-            fileName = self.fileList[file_idx]
-            try:
-                # Try to find the h5 data file associated with this iteration
-                dataIn = h5py.File(fileName, "r")
-                dsetName = "/iterations/iter_%08d/seg_index" % int(n_iter)
-
-                # Check if the dataset
-                dataset_exists = dsetName in dataIn
-
-                # Check to make sure this isn't the last iteration -- last iterations have incomplete data
-                is_not_last_iteration = (
-                    "/iterations/iter_%08d/seg_index" % int(n_iter + 1) in dataIn
-                )
-
-                log.debug(f"From file {fileName}, loading iteration {n_iter}")
-
-                if dataset_exists and is_not_last_iteration:
-
-                    dset = dataIn[dsetName]
-                    newSet = dset[:]
-                    n_segs_in_file = np.shape(newSet)
-                    n_segs_in_file = n_segs_in_file[0]
-                    dsetNameP = "/iterations/iter_%08d/pcoord" % int(n_iter)
-                    dsetP = dataIn[dsetNameP]
-                    pcoord = dsetP[:]
-                    weights = dset["weight"]
-                    seg_weights = np.append(seg_weights, weights)
-
-                    if (
-                        not pcoord.shape[2] == self.pcoord_ndim
-                        and not self.pcoord_shape_warned
-                    ):
-                        log.warning(
-                            f"Dimensions of pcoord in {fileName} ({pcoord.shape[2]}) do not match specified "
-                            f"pcoord dimensionality self.pcoord_ndim ({self.pcoord_ndim}). "
-                            f"MSM-WE will only load up to dimension {self.pcoord_ndim}. "
-                            f"This pcoord is just used for stratification, and this is expected behavior if "
-                            f"you're extending your pcoord (i.e. in an optimization flow). "
-                        )
-                        self.pcoord_shape_warned = True
-
-                    # Iterate over segments in this dataset
-                    for seg_idx in range(n_segs_in_file):
-                        # if np.sum(pcoord[seg_idx,self.pcoord_len-1,:])==0.0:
-                        # # intentionally using this to write in dummy pcoords,
-                        # # this is a good thing to have for post-analysis though!
-                        #    raise ValueError('Sum pcoord is 0, probably middle of WE iteration, not using iteration') f
-                        westList = np.append(westList, file_idx)
-                        segindList = np.append(segindList, seg_idx)
-                        weightList = np.append(weightList, newSet[seg_idx][0])
-                        pcoord0List = np.append(
-                            pcoord0List,
-                            np.expand_dims(pcoord[seg_idx, 0, : self.pcoord_ndim], 0),
-                            axis=0,
-                        )
-                        pcoord1List = np.append(
-                            pcoord1List,
-                            np.expand_dims(
-                                pcoord[
-                                    seg_idx, self.pcoord_len - 1, : self.pcoord_ndim
-                                ],
-                                0,
-                            ),
-                            axis=0,
-                        )
-                        n_segs = n_segs + 1
-                dataIn.close()
-            except Exception as dataset_exists:
-                sys.stdout.write("error in " + fileName + str(sys.exc_info()[0]) + "\n")
-                raise dataset_exists
-
-        # log.debug(f"Found {n_segs} segments in iteration {n_iter}")
-
-        self.westList = westList.astype(int)
-
-        # This is a list of the segment indices
-        self.segindList = segindList.astype(int)
-        self.seg_weights[n_iter] = seg_weights
-        self.weightList = weightList
-        self.nSeg = n_segs
-        self.pcoord0List = pcoord0List
-        self.pcoord1List = pcoord1List
-
-    def get_iterations(self):
-        """
-        Updates internal state with the maximum number of iterations, and the number of segments in each section.
-
-        Note
-        ----
-        This updates :code:`numSegments` -- :code:`numSegments` is actually a *list* of the number of segments in each iteration.
-
-        Returns
-        -------
-        None
-        """
-
-        log.debug("Getting number of iterations and segments")
-
-        numSegments = np.array([])
-        nSeg = 1
-        n_iter = 1
-
-        # Loop over nSegs
-        # TODO: Not sure I understand the logic in this loop
-        while nSeg > 0:
-            nSeg = 0
-
-            # Iterate through each filename in fileList, and see if it contains the iteration we're looking for
-            # TODO: This loop is pretty common, this should be refactored into a find_iteration() or something
-            for file_index in range(self.n_data_files):
-                fileName = self.fileList[file_index]
-                try:
-                    dataIn = h5py.File(fileName, "r")
-                    dsetName = "/iterations/iter_%08d/seg_index" % int(n_iter)
-                    dataset_exists = dsetName in dataIn
-
-                    is_not_last_iteration = (
-                        "/iterations/iter_%08d/seg_index" % int(n_iter + 1) in dataIn
-                    )
-
-                    if dataset_exists and is_not_last_iteration:
-                        # If this file does contain the iteration of interest
-                        # if dataset_exists:
-                        dset = dataIn[dsetName]
-                        newSet = dset[:]
-                        nS = np.shape(newSet)
-                        nSeg = nS[0] + nSeg
-                    dataIn.close()
-                except Exception as e:
-                    log.error(e)
-                    log.error(f"No segments in {fileName} {str(sys.exc_info()[0])}")
-
-            if nSeg > 0:
-                numSegments = np.append(numSegments, nSeg)
-                log.debug(
-                    "Iteration " + str(n_iter) + " has " + str(nSeg) + " segments...\n"
-                )
-
-            n_iter = n_iter + 1
-
-        # Warning: These are not defined until this is run for the first time
-        self.numSegments = numSegments
-        self.maxIter = numSegments.size
-
-    def get_iterations_iters(self, first_iter: int, last_iter: int):
-        """
-        Updates internal state with the maximum number of iterations, and the number of segments in each section.
-
-        Parameters
-        ----------
-        first_iter : int
-        last_iter : int
-
-        Returns
-        -------
-        None
-
-        Warning
-        ----
-        This is potentially deprecated or unnecessary. Currently unused.
-
-        """
-
-        numSegments = np.array([])
-
-        for n_iter in range(first_iter, last_iter + 1):
-            nSeg = 0
-            for iF in range(self.n_data_files):
-                fileName = self.fileList[iF]
-                try:
-                    dataIn = h5py.File(fileName, "r")
-                    dsetName = "/iterations/iter_%08d/seg_index" % int(n_iter)
-                    dataset_exists = dsetName in dataIn
-                    if dataset_exists:
-                        dset = dataIn[dsetName]
-                        newSet = dset[:]
-                        nS = np.shape(newSet)
-                        nSeg = nS[0] + nSeg
-                    dataIn.close()
-                except Exception as e:
-                    log.error(e)
-                    log.error(f"No segments in {fileName} {str(sys.exc_info()[0])}")
-
-            if nSeg > 0:
-                numSegments = np.append(numSegments, nSeg)
-                sys.stdout.write(
-                    "Iteration " + str(n_iter) + " has " + str(nSeg) + " segments...\n"
-                )
-        self.numSegments = numSegments
-        self.maxIter = last_iter
 
     def set_topology(self, topology):
         """
@@ -1410,1051 +1161,6 @@ class modelWE:
         else:
             raise NotImplementedError("Unsupported topology")
 
-    def get_transition_data(self, n_lag):
-        """
-        This function analyzes pairs of coordinates at the current iteration, set by :code:`self.n_iter`, and at some
-            lag in the past, :code:`self.n_iter - n_lag`.
-
-        Segments where a walker was warped (recycled) use the basis coords as the lagged coords.
-
-        Parameters
-        ----------
-        n_lag : int
-            Number of lags to use for transitions.
-
-        Returns
-        -------
-        None
-        """
-
-        log.warning(
-            "Getting transition data at arbitrary lags > 0 is not yet supported! Use at your own risk."
-        )
-
-        # get segment history data at lag time n_lag from current iter
-        if n_lag > self.n_iter:
-            sys.stdout.write(
-                "too much lag for iter... n_lag reduced to: " + str(self.n_iter) + "\n"
-            )
-            n_lag = self.n_iter
-        if n_lag >= self.n_hist:
-            sys.stdout.write("too much lag for stored history... recalculating...\n")
-            self.get_seg_histories(n_lag)
-
-        self.n_lag = n_lag
-        segindList_lagged = self.seg_histories[:, n_lag]
-
-        # TODO: What exactly is this a list of?
-        # seg_histories is a list of indices of the segments
-        warpList = self.seg_histories[:, 0:n_lag]  # check for warps
-        warpList = np.sum(warpList < 0, 1)
-
-        # Get the weights for the lagged and current iterations
-        # If something was split/merged between n_iter and n_iter-n_lag , then the weights may have changed, so
-        #   check a particular segment's weight.
-        # TODO: Does this effectively get the parent weight if it was split from something else? Then weight_histories
-        #    would need to be tracking parents/children
-        weightList_lagged = self.weight_histories[:, n_lag]
-        # TODO: Does this copy need to be made?
-        weightList = self.weightList
-
-        # This will become a list of (lagged iter coord, current iter coord)
-        coordPairList = np.zeros((self.nSeg, self.nAtoms, self.coord_ndim, 2))
-
-        prewarpedStructures = np.zeros((self.nSeg, self.nAtoms, self.coord_ndim))
-        nWarped = 0
-
-        # Go through each segment, and get pairs of coordinates at current iter (n_iter) and
-        # lagged iter (n_iter-n_lag)
-        for seg_idx in range(self.nSeg):
-
-            # FIXME: Try statements should encompass the smallest amount of code
-            #  possible - anything could be tripping this
-            # try:
-            if seg_idx == 0:
-                westFile = self.fileList[self.westList[seg_idx]]
-                dataIn = h5py.File(westFile, "r")
-                dsetName = "/iterations/iter_%08d/auxdata/%s" % (
-                    int(self.n_iter),
-                    self.auxpath,
-                )
-                dset = dataIn[dsetName]
-                coords_current = dset[:]
-                dsetName = "/iterations/iter_%08d/auxdata/%s" % (
-                    int(self.n_iter - n_lag),
-                    self.auxpath,
-                )
-                dset = dataIn[dsetName]
-                coords_lagged = dset[:]
-            elif self.westList[seg_idx] != self.westList[seg_idx - 1]:
-                # FIXME: I think you can just move this close to an if statement in the beginning, and then remove
-                #   this whole if/elif. Everything after that close() seems to be duplicated.
-                dataIn.close()
-                westFile = self.fileList[self.westList[seg_idx]]
-                dataIn = h5py.File(westFile, "r")
-
-                # Load the data for the current iteration
-                dsetName = "/iterations/iter_%08d/auxdata/%s" % (
-                    int(self.n_iter),
-                    self.auxpath,
-                )
-                dset = dataIn[dsetName]
-                coords_current = dset[:]
-
-                # Load the lagged data for (iteration - n_lag)
-                dsetName = "/iterations/iter_%08d/auxdata/%s" % (
-                    int(self.n_iter - n_lag),
-                    self.auxpath,
-                )
-                dset = dataIn[dsetName]
-                coords_lagged = dset[:]
-
-            coordPairList[seg_idx, :, :, 1] = coords_current[
-                self.segindList[seg_idx], 1, :, :
-            ]
-
-            # If this segment has no warps, then add the lagged coordinates
-            if warpList[seg_idx] == 0:
-                # Try to set the previous coord in the transition pair to the segment's lagged coordinates
-                try:
-                    lagged_seg_index = segindList_lagged[seg_idx]
-                    coordPairList[seg_idx, :, :, 0] = coords_lagged[
-                        lagged_seg_index, 0, :, :
-                    ]
-
-                # If this fails, then there were no lagged coordinates for this structure.
-                except IndexError as e:
-                    log.critical(
-                        f"Lagged coordinates do not exist for the structure in segment {seg_idx}"
-                    )
-                    raise e
-                    weightList_lagged[seg_idx] = 0.0
-                    weightList[
-                        seg_idx
-                    ] = 0.0  # set transitions without structures to zero weight
-
-            # If something was recycled during this segment, then instead of using the lagged cooordinates,
-            #   just use the basis coords.
-            # But, also save the original structure before the warp!
-            elif warpList[seg_idx] > 0:
-
-                # St
-                prewarpedStructures[nWarped, :, :] = coords_lagged[
-                    segindList_lagged[seg_idx], 0, :, :
-                ]
-
-                assert self.basis_coords is not None
-
-                coordPairList[seg_idx, :, :, 0] = self.basis_coords
-                nWarped = nWarped + 1
-
-            # # TODO: What triggers this? When that critical log hits, come update this comment and the main docstring.
-            # # This triggers on index out of bounds... But that's a huge try statement!
-            # except Exception as e:
-            #     log.critical("Document whatever's causing this exception!")
-            #     log.warning(e)
-            #     raise e
-            #     weightList_lagged[seg_idx] = 0.0
-            #     weightList[
-            #         seg_idx
-            #     ] = 0.0  # set transitions without structures to zero weight
-
-        # Squeeze removes axes of length 1 -- this helps with np.where returning nested lists of indices
-        # FIXME: is any of this necessary? This kinda seems like it could be replaced with
-        #  something like indWarped = np.squeeze(np.where(warpList > 0)).astype(int)
-        indWarped = np.squeeze(np.where(warpList > 0))
-        indWarpedArray = np.empty(nWarped)
-        indWarpedArray[0:nWarped] = indWarped
-        indWarped = indWarpedArray.astype(int)
-
-        # Get the current and lagged weights
-        # This returns (the wrong? none at all?) weights for segments with warps, which is corrected below
-        transitionWeights = weightList.copy()
-        departureWeights = weightList_lagged.copy()
-
-        # Get correct weights for segments that warped
-        for iW in range(nWarped):
-
-            # The coord pair here is (pre-warped structure, reference topology) instead of
-            #   (lagged struture, current structure)
-            coordPair = np.zeros((1, self.nAtoms, self.coord_ndim, 2))
-            coordPair[0, :, :, 0] = prewarpedStructures[iW, :, :]
-            coordPair[0, :, :, 1] = self.reference_coord
-            coordPairList = np.append(coordPairList, coordPair, axis=0)
-
-            # TODO: iterWarped appears to be the iteration the warp happened at
-            iterWarped = np.squeeze(np.where(self.seg_histories[indWarped[iW], :] < 0))
-            try:
-                nW = np.shape(iterWarped)
-                iterWarped = iterWarped[0]
-                sys.stdout.write(
-                    "    segment "
-                    + str(indWarped[iW])
-                    + " warped "
-                    + str(nW)
-                    + " times\n"
-                )
-            # TODO: What exception is this handling? Why would this be tripped?
-            except Exception as e:
-                log.critical(f"Exception was {e}")
-                log.critical("UPDATE THIS EXCEPTION HANDLER!")
-
-                sys.stdout.write(
-                    "    segment " + str(indWarped[iW]) + " warped 1 time\n"
-                )
-
-            # The "current" weights are taken at the warp iteration, so the final pair of weights are
-            #   (the lagged weights, the weights at the warp)
-            transitionWeights = np.append(
-                transitionWeights, self.weight_histories[indWarped[iW], iterWarped]
-            )
-            departureWeights = np.append(
-                departureWeights, self.weight_histories[indWarped[iW], n_lag]
-            )
-
-        self.coordPairList = coordPairList
-        self.transitionWeights = transitionWeights
-        self.departureWeights = departureWeights
-
-    def get_transition_data_lag0(self):
-        """
-        Get coordinate pairs at the beginning and end of this iteration.
-
-        Updates:
-            - self.coordPairList, a list of  parent/child coordinate pairs
-            - self.transitionWeights, a copy of self.weightList
-            - self.departureWeights, a copy of self.weightList
-        """
-
-        weightList = self.weightList
-        coordPairList = np.zeros((self.nSeg, self.nAtoms, self.coord_ndim, 2))
-
-        log.debug(
-            f"Getting transition data for {self.nSeg} segs in iteration {self.n_iter}, at a lag of 0"
-        )
-
-        # the segments in this iteration may be split across a number of different files
-        # "Traditionally", we store a reference for  each segment  of which WEST file it's in
-        # But what if we flip that, and for each west file, get which segments are in it?
-
-        seg_west_files = self.westList[range(self.nSeg)]
-        west_file_idxs = np.unique(seg_west_files)
-        west_files = [self.fileList[idx] for idx in west_file_idxs]
-
-        for idx, west_file in enumerate(west_files):
-
-            segs_contained = np.where(seg_west_files == idx)[0]
-
-            with h5py.File(west_file, "r") as data_file:
-
-                dsetName = "/iterations/iter_%08d/auxdata/%s" % (
-                    int(self.n_iter),
-                    self.auxpath,
-                )
-
-                try:
-                    dset = data_file[dsetName]
-                except KeyError as e:
-                    raise e
-
-                coords = dset
-
-                coordPairList[segs_contained, :, :, 0] = coords[:, 0, :, :]
-                coordPairList[segs_contained, :, :, 1] = coords[
-                    :, self.pcoord_len - 1, :, :
-                ]
-
-                # Check for NaNs in segments
-                nan_segments = segs_contained[
-                    np.argwhere(
-                        np.isnan(coordPairList[segs_contained]).any(axis=(1, 2, 3))
-                    )
-                ]
-
-                if nan_segments.shape[0] > 0:
-                    log.warning(
-                        f"Bad coordinates for segments {nan_segments}, setting weights to 0"
-                    )
-                    weightList[nan_segments] = 0.0
-
-        transitionWeights = weightList.copy()
-        departureWeights = weightList.copy()
-
-        self.coordPairList = coordPairList
-        self.transitionWeights = transitionWeights
-        self.departureWeights = departureWeights
-
-    def get_seg_histories(self, n_hist):
-        """
-        **TODO: What does this do exactly?**
-
-        Updates:
-            - self.seg_histories
-            - self.weight_histories
-            - self.n_hist
-
-        Parameters
-        ----------
-        n_hist : int
-            Number of steps of history information to include.
-
-        Returns
-        -------
-        None
-
-        """
-
-        log.debug(f"Getting seg histories for {self.nSeg} segments")
-
-        if n_hist > self.n_iter:
-            sys.stdout.write(
-                "we have too much history... n_hist reduced to: "
-                + str(self.n_iter)
-                + "\n"
-            )
-            n_hist = self.n_iter
-
-        # FIXME: The only other place this is used is in get_transition_data(). May need to see how it's handled there
-        #   i.e. make sure it's initialized
-        self.n_hist = n_hist
-
-        seg_histories = np.zeros((self.nSeg, self.n_hist + 1))
-        weight_histories = np.zeros((self.nSeg, self.n_hist))
-
-        # Loop over all segments
-        for iS in range(self.nSeg):
-
-            # Print a message every 100 segments
-            # if iS % 100 == 0:
-            #     sys.stdout.write(
-            #         "        getting history for iteration "
-            #         + str(self.n_iter)
-            #         + " segment "
-            #         + str(iS)
-            #         + "...\n"
-            #     )
-
-            # Open the relevant datafile for reading
-            if iS == 0:
-                westFile = self.fileList[self.westList[iS]]
-                dataIn = h5py.File(westFile, "r")
-            elif self.westList[iS] != self.westList[iS - 1]:
-                # If you're past the 0th segment, then close the previous file.
-                # FIXME: Close the file after finished working with it, at the *end* of the loop. As-is, the last loop
-                #       leaves the file open.
-                dataIn.close()
-                westFile = self.fileList[self.westList[iS]]
-                dataIn = h5py.File(westFile, "r")
-
-            seg_histories[iS, 0] = self.segindList[iS]
-            warped = 0
-
-            # Iterate over history lengths from 1 to n_hist
-            for iH in range(1, self.n_hist + 1):
-
-                indCurrentSeg = seg_histories[iS, iH - 1]
-
-                if indCurrentSeg < 0 and warped == 0:
-                    sys.stdout.write(
-                        "Segment "
-                        + str(iS)
-                        + " warped last iter: History must end NOW!\n"
-                    )
-                    warped = 1
-
-                elif indCurrentSeg >= 0 and warped == 0:
-                    dsetName = "/iterations/iter_%08d/seg_index" % int(
-                        self.n_iter - iH + 1
-                    )
-
-                    dset = dataIn[dsetName]
-                    seg_histories[iS, iH] = dset[indCurrentSeg][1]
-                    weight_histories[iS, iH - 1] = dset[indCurrentSeg][0]
-
-                    if seg_histories[iS, iH] < 0:
-                        sys.stdout.write(
-                            "            segment "
-                            + str(iS)
-                            + " warped "
-                            + str(iH)
-                            + " iters ago\n"
-                        )
-        self.seg_histories = seg_histories[:, :-1].astype(int)
-        self.weight_histories = weight_histories
-        # FIXME:
-        #   weight histories and segment histories go in reverse order,
-        #   so final current iter is first of 0 index
-
-    def collect_iter_coordinates(self):  # grab coordinates from WE traj_segs folder
-        """
-        Goes through the generated trajectory segments, and adds data from the segments to an H5 file.
-
-        This should be implemented by the user, and this implementation assumes a really specific configuration.
-        This is left in mostly as an example.
-
-        Returns
-        -------
-        None
-
-        Todo
-        ----
-        Generalize to different filetypes. This appears to be AMBER specific and relies on loading rst7 files
-        """
-
-        log.critical(
-            "If you're calling this function, be absolutely sure it does what you want -- this is "
-            "very specifically written for the output of certain simulations."
-        )
-
-        nS = self.nSeg
-        westFile = self.fileList[self.westList[0]]
-        dataIn = h5py.File(westFile, "a")
-        coords = np.zeros((0, 2, self.nAtoms, self.coord_ndim))
-        for iS in range(self.nSeg):
-            # FIXME: Replace strings with Pathlib paths
-            westFile = self.fileList[self.westList[iS]]
-            WEfolder = westFile.replace("west.h5", "")
-            trajpath = WEfolder + "traj_segs/%06d/%06d" % (self.n_iter, iS)
-            coord0 = np.squeeze(
-                md.load(
-                    trajpath + "/parent.rst7", top=self.reference_structure.topology
-                )._xyz
-            )
-            coord1 = np.squeeze(
-                md.load(
-                    trajpath + "/seg.rst7", top=self.reference_structure.topology
-                )._xyz
-            )
-            coordT = np.array([coord0, coord1])
-            coordT = coordT[np.newaxis, :, :, :]
-            coords = np.append(coords, coordT, axis=0)
-            try:
-                if iS > 0:
-                    if self.westList[iS] != self.westList[iS - 1] and iS < nS - 1:
-                        dsetName = "/iterations/iter_%08d/auxdata/%s" % (
-                            int(self.n_iter),
-                            self.auxpath,
-                        )
-                        try:
-                            # TODO: Why exclude the last point?
-                            dset = dataIn.create_dataset(
-                                dsetName, np.shape(coords[:-1, :, :, :])
-                            )
-                            dset[:] = coords[:-1, :, :, :]
-                        except (RuntimeError, ValueError):
-                            del dataIn[dsetName]
-                            dset = dataIn.create_dataset(
-                                dsetName, np.shape(coords[:-1, :, :, :])
-                            )
-                            dset[:] = coords[:-1, :, :, :]
-                            log.warning(
-                                "coords exist for iteration "
-                                + str(self.n_iter)
-                                + " overwritten\n"
-                            )
-
-                        dataIn.close()
-                        coords = np.zeros((0, 2, self.nAtoms, self.coord_ndim))
-                        coords = np.append(coords, coordT, axis=0)
-                        dataIn = h5py.File(westFile, "a")
-
-                    # If it's the last segment, don't exclude the last point (why?)
-                    elif iS == nS - 1:
-                        dsetName = "/iterations/iter_%08d/auxdata/%s" % (
-                            int(self.n_iter),
-                            self.auxpath,
-                        )
-                        try:
-                            dset = dataIn.create_dataset(dsetName, np.shape(coords))
-                            dset[:] = coords
-                        except (RuntimeError, ValueError):
-                            del dataIn[dsetName]
-                            dset = dataIn.create_dataset(dsetName, np.shape(coords))
-                            dset[:] = coords
-                            log.warning(
-                                "coords exist for iteration "
-                                + str(self.n_iter)
-                                + " overwritten\n"
-                            )
-
-                        dataIn.close()
-
-            except Exception as e:
-                log.error(
-                    "error collecting coordinates from "
-                    + WEfolder
-                    + " , iter "
-                    + str(self.n_iter)
-                    + " segment "
-                    + str(self.segindList[iS])
-                    + "\n"
-                )
-                log.error(e)
-
-    def get_iter_coordinates(self, iteration):
-        """
-        Return the valid coordinates for a certain iteration
-
-        Parameters
-        ----------
-        iteration: int
-            The iteration to return coordinates for
-
-        Returns
-        -------
-        Array of coordinates for all atoms at the current iteration
-        """
-
-        self.load_iter_data(iteration)
-        self.load_iter_coordinates()
-        # indGood = np.squeeze(np.where(np.sum(np.sum(self.cur_iter_coords, 2), 1) != 0))
-
-        bad_coords = np.isnan(self.cur_iter_coords).any(axis=(1, 2))
-        good_coords = ~bad_coords
-
-        # iter_coords = self.cur_iter_coords[indGood]
-        iter_coords = self.cur_iter_coords[good_coords]
-
-        return iter_coords
-
-    def load_iter_coordinates(self):
-        """
-
-        Returns
-        -------
-
-        """
-
-        cur_iter_coords = np.full(
-            (self.nSeg, self.nAtoms, self.coord_ndim), fill_value=np.nan
-        )
-
-        log.debug(
-            f"Getting coordinates for {self.nSeg} segs in iteration {self.n_iter}, at a lag of 0"
-        )
-
-        # the segments in this iteration may be split across a number of different files
-        # "Traditionally", we store a reference for  each segment  of which WEST file it's in
-        # But flip that, and for each west file, get which segments are in it
-
-        seg_west_files = self.westList[range(self.nSeg)]
-        west_file_idxs = np.unique(seg_west_files)
-        west_files = [self.fileList[idx] for idx in west_file_idxs]
-
-        for idx, west_file in enumerate(west_files):
-
-            segs_contained = np.where(seg_west_files == idx)[0]
-
-            with h5py.File(west_file, "r") as data_file:
-
-                dsetName = "/iterations/iter_%08d/auxdata/%s" % (
-                    int(self.n_iter),
-                    self.auxpath,
-                )
-
-                try:
-                    dset = data_file[dsetName]
-                except (RuntimeError, ValueError, KeyError):
-                    log.error(
-                        f"Error getting coordinates from {west_file}, in iteration {self.n_iter}"
-                    )
-
-                    cur_iter_coords[segs_contained, :, :] = np.full(
-                        (self.nAtoms, self.coord_ndim), fill_value=np.nan
-                    )
-                    self.coordsExist = False
-
-                    # raise e
-
-                else:
-                    coords = dset
-                    cur_iter_coords[segs_contained, :, :] = coords[:, 1, :, :]
-
-        self.cur_iter_coords = cur_iter_coords
-
-    def load_iter_coordinates0(self):  # get iteration initial coordinates
-        coordList = np.full(
-            (self.nSeg, self.nAtoms, self.coord_ndim), fill_value=np.nan
-        )
-        for iS in range(self.nSeg):
-            if iS == 0:
-                westFile = self.fileList[self.westList[iS]]
-                dataIn = h5py.File(westFile, "r")
-                dsetName = "/iterations/iter_%08d/auxdata/%s" % (
-                    int(self.n_iter),
-                    self.auxpath,
-                )
-                dset = dataIn[dsetName]
-                coord = dset[:]
-            elif self.westList[iS] != self.westList[iS - 1]:
-                dataIn.close()
-                westFile = self.fileList[self.westList[iS]]
-                dataIn = h5py.File(westFile, "r")
-                dsetName = "/iterations/iter_%08d/auxdata/%s" % (
-                    int(self.n_iter),
-                    self.auxpath,
-                )
-                dset = dataIn[dsetName]
-                coord = dset[:]
-            coordList[iS, :, :] = coord[self.segindList[iS], 0, :, :]
-        self.cur_iter_coords = coordList
-
-    def get_coordinates(self, first_iter, last_iter):
-        """
-        Unused
-
-        Parameters
-        ----------
-        first_iter
-        last_iter
-
-        Returns
-        -------
-        """
-
-        log.warning("This function is not tested or supported, use at your own risk!")
-
-        self.first_iter = first_iter
-        self.last_iter = last_iter
-        iters = range(self.first_iter, self.last_iter + 1)
-        coordSet = np.zeros((0, self.nAtoms, self.coord_ndim))
-        for iter in iters:
-            if iter % 50 == 0:
-                sys.stdout.write(
-                    "    gathering structures from iteration " + str(iter) + "...\n"
-                )
-            self.load_iter_data(iter)
-            self.load_iter_coordinates()
-            if self.coordsExist:
-                coordSet = np.append(coordSet, self.cur_iter_coords, axis=0)
-        self.all_coords = coordSet
-
-    def get_coordSet(self, last_iter, streaming=None):
-        """
-        Loads all coordinates and progress coordinates into memory for later usage.
-
-        If streaming, then this only loads pcoords
-
-        Todo
-        ----
-        I want to avoid loading full coordinates into memory as much as possible.
-        That means trying to replace usage of all_coords here.
-
-        Parameters
-        ----------
-        last_iter
-
-        Returns
-        -------
-
-        """
-
-        if streaming is None and False:  # and self.dimReduceMethod == "vamp":
-            streaming = False
-        elif streaming is None:
-            streaming = True
-
-        log.debug(f"CoordSet loading up to {last_iter}")
-        total_segments = int(sum(self.numSegments[:last_iter]))
-
-        if not streaming:
-            coordSet = np.full(
-                (total_segments, self.nAtoms, self.coord_ndim), fill_value=np.nan
-            )
-        pcoordSet = np.full((total_segments, self.pcoord_ndim), fill_value=np.nan)
-
-        last_seg_idx = total_segments
-
-        # Update iterations N+1 -> 1
-        for i in tqdm.tqdm(range(last_iter, 0, -1), desc="Getting coordSet"):
-
-            self.load_iter_data(i)
-            self.load_iter_coordinates()
-
-            first_seg_idx = last_seg_idx - len(self.segindList)
-            assert first_seg_idx >= 0, "Referencing a segment that doesn't exist"
-
-            # Get the indices of all "good" coordinates, where a  valid coordinate has been obtained
-            # Boolean mask for coords that have been successfully initialized
-            bad_coords = np.isnan(self.cur_iter_coords).any(axis=(1, 2))
-            good_coords = ~bad_coords
-
-            if bad_coords.any():
-                log.warning(
-                    f"Walker {np.argwhere(bad_coords).squeeze()} has bad coords in iteration(s) {self.n_iter}"
-                )
-
-            if not streaming:
-                coordSet[first_seg_idx:last_seg_idx][
-                    good_coords
-                ] = self.cur_iter_coords[good_coords, :, :]
-
-            pcoordSet[first_seg_idx:last_seg_idx][good_coords] = self.pcoord1List[
-                good_coords, :
-            ]
-
-            last_seg_idx = first_seg_idx
-
-        # Set the coords, and pcoords
-        if not streaming:
-            self.all_coords = coordSet
-        else:
-            pass
-
-        self.pcoordSet = pcoordSet
-
-        first_iter_cluster = i
-        self.first_iter = first_iter_cluster
-        self.last_iter = last_iter
-
-    def get_traj_coordinates(self, from_iter, traj_length):
-        if traj_length > from_iter:
-            traj_length = from_iter - 1
-            sys.stdout.write(
-                "trajectory length too long: set to " + str(traj_length) + "\n"
-            )
-        self.load_iter_data(from_iter)
-        self.get_seg_histories(traj_length)
-        traj_iters = np.zeros((traj_length, self.nSeg, self.nAtoms, self.coord_ndim))
-        ic = traj_length - 1
-        iH = 0
-        nS = self.nSeg
-        westList = self.westList.copy()
-        for i in range(from_iter, from_iter - traj_length, -1):
-            sys.stdout.write(
-                "    gathering structures from iteration " + str(i) + "...\n"
-            )
-            self.load_iter_data(i)
-            self.load_iter_coordinates()
-            seg_history = self.seg_histories[:, iH]
-            seg_history_index = np.zeros(nS).astype(int)
-            last_index = np.zeros(nS).astype(
-                int
-            )  # list of final iterations for warped particles
-            for iS in range(nS):
-                if seg_history[iS] > 0:
-                    seg_history_index[iS] = np.where(
-                        np.logical_and(
-                            self.segindList == seg_history[iS],
-                            self.westList == westList[iS],
-                        )
-                    )[0][
-                        0
-                    ]  # segment indexes are local to the westfile, not global for the analysis set
-                    last_index[iS] = 0
-                elif seg_history[iS] < 0:
-                    last_index[iS] = i - (from_iter - traj_length) - 1
-            traj_iters[ic, :, :, :] = self.cur_iter_coords[seg_history_index, :, :]
-            ic = ic - 1
-            iH = iH + 1
-        self.load_iter_data(from_iter)
-        traj_iters = np.swapaxes(traj_iters, 0, 1)
-        self.trajSet = [None] * self.nSeg
-        for iS in range(self.nSeg):
-            self.trajSet[iS] = traj_iters[iS, last_index[iS] :, :, :]
-
-    def do_pca(self, arg):
-        rough_pca, iteration, processCoordinates = arg
-        iter_coords = self.get_iter_coordinates(iteration)
-
-        # If  no good coords in this iteration, skip it
-        if iter_coords.shape[0] == 0:
-            return rough_pca
-
-        processed_iter_coords = processCoordinates(iter_coords)
-        rough_pca.partial_fit(processed_iter_coords)
-
-        log.debug(f"{rough_pca.n_samples_seen_} samples seen")
-
-        return rough_pca
-
-    def do_full_pca(self, arg):
-
-        ipca, iteration, processCoordinates, components_for_var = arg
-
-        iter_coords = self.get_iter_coordinates(iteration)
-
-        used_iters = 0
-        # Keep adding coords until you have more than your components
-        while iter_coords.shape[0] <= components_for_var:
-
-            used_iters += 1
-
-            _iter_coords = self.get_iter_coordinates(iteration + used_iters)
-            if _iter_coords.shape[0] == 0:
-                continue
-
-            iter_coords = np.append(iter_coords, _iter_coords, axis=0)
-
-        processed_iter_coords = processCoordinates(iter_coords)
-        log.debug(
-            f"About to run iPCA on  {processed_iter_coords.shape} processed coords"
-        )
-        ipca.partial_fit(processed_iter_coords)
-
-        return ipca, used_iters
-
-    def dimReduce(
-        self,
-        first_iter=1,
-        first_rough_iter=None,
-        last_iter=None,
-        rough_stride=10,
-        fine_stride=1,
-        variance_cutoff=0.95,
-        use_weights=True,
-    ):
-        """
-        Dimensionality reduction using the scheme specified in initialization.
-
-        This just defines the dimensionality reduction scheme and builds the model -- it does NOT actually transform
-        the data!
-
-        Transforming the data is performed via reduceCoordinates(), which uses self.coordinates as set
-            by this.
-
-        Updates:
-            - `self.coordinates`
-            - `self.ndim`
-
-        Returns
-        -------
-        None
-        """
-
-        log.debug(f"Running dimensionality reduction -- method: {self.dimReduceMethod}")
-
-        # log.debug(self.coordSet)
-        if self.dimReduceMethod == "pca":
-
-            # Do this in a streaming way, iteration by iteration
-            # First, do a "rough" PCA on the last 10% of the data to get the number of components that explain the
-            #   variance cutoff.
-            # This is necessary because with incremental PCA, there's no way to do this ahead of time.
-
-            if last_iter is None:
-                last_iter = self.maxIter
-
-            rough_ipca = iPCA()
-
-            if first_rough_iter is None:
-                total_num_iterations = last_iter
-                # Stride every 10th frame, so you're only doing the "rough" pca on 10% of the data
-                if total_num_iterations > 100:
-                    rough_iters = range(1, total_num_iterations, rough_stride)
-
-                # But if you only have 100 frames or fewer, then just do the last-half.
-                else:
-                    rough_iters = range(total_num_iterations // 2, total_num_iterations)
-            else:
-                rough_iters = range(first_rough_iter, last_iter, rough_stride)
-
-            for iteration in tqdm.tqdm(rough_iters, desc="Initial iPCA"):
-
-                # TODO: Allow  chunking here so you don't have  to  go 1  by  1, but N by N
-                # If you don't use 'fork' context here, this will break in Jupyter.
-                # That's because processCoordinates is monkey-patched in. With 'spawn' (i.e. without fork), the module
-                #   is re-imported in the child process. In the reimported  module, processCoordinates is undefined.
-                # With 'fork', it preserves the monkey-patched version.
-                # Additionally, 'fork' is a little faster than  spawn. Ironically, that's usually at the cost  of memory
-                #   usage. But here, the memory being used by the main thread (and therefore being copied here) isn't
-                #   that great -- the memory issue stems from it not being freed up between successive calls.
-                with concurrent.futures.ProcessPoolExecutor(
-                    max_workers=1, mp_context=mp.get_context("fork")
-                ) as executor:
-                    rough_ipca = executor.submit(
-                        self.do_pca, [rough_ipca, iteration, self.processCoordinates]
-                    ).result()
-
-            components_for_var = (
-                np.argmax(
-                    np.cumsum(rough_ipca.explained_variance_ratio_) > variance_cutoff
-                )
-                + 1
-            )
-            log.debug(f"Keeping {components_for_var} components")
-            components_for_var = min(
-                components_for_var, (last_iter - first_iter) // fine_stride - 1
-            )
-
-            # Now do the PCA again, with that many components, using all the iterations.
-            ipca = iPCA(n_components=components_for_var)
-
-            extra_iters_used = 0
-            for iteration in tqdm.tqdm(
-                range(first_iter, last_iter, fine_stride), desc="iPCA"
-            ):
-
-                if extra_iters_used > 0:
-                    extra_iters_used -= 1
-                    log.debug(f"Already processed  iter  {iteration}")
-                    continue
-
-                # Try some stuff to help memory management. I think  a lot of memory is not being explicitly released
-                #   here when I'm looping, because I can watch my swap usage steadily grow while it's running this loop.
-                # https://stackoverflow.com/questions/1316767/how-can-i-explicitly-free-memory-in-python has some good
-                #   details  on how memory may be freed by Python, but not necessarily recognized  as free by the OS.
-                # One "guaranteed" way to free  memory back to the OS that's been released by Python is to do  the memory
-                #   intensive operation  in a subprocess. So, maybe I need to do my partial fit in a subprocess.
-                # In fact, I first moved partial_fit alone to a subprocess, but that didn't help. The issue isn't
-                #   partial_fit, it's actually loading the coords.
-                with concurrent.futures.ProcessPoolExecutor(
-                    max_workers=1, mp_context=mp.get_context("fork")
-                ) as executor:
-                    ipca, extra_iters_used = executor.submit(
-                        self.do_full_pca,
-                        [ipca, iteration, self.processCoordinates, components_for_var],
-                    ).result()
-
-            self.coordinates = ipca
-            self.ndim = components_for_var
-
-        # elif self.dimReduceMethod == "vamp":
-        #     # TODO: I  don't think trajSet is initialized by itself -- you need to manually call get_traj_coordinates
-        #     log.warning(
-        #         "VAMP dimensionality reduction requires you to *manually* call get_traj_coordinates first, "
-        #         "or self.trajSet will be all None. Make sure you've done that!"
-        #     )
-        #     raise NotImplementedError
-        #
-        #     ntraj = len(self.trajSet)
-        #     data = [None] * ntraj
-        #     for itraj in range(ntraj):
-        #         data[itraj] = self.processCoordinates(self.trajSet[itraj])
-        #     self.coordinates = coor.vamp(
-        #         data,
-        #         lag=self.vamp_lag,
-        #         dim=self.vamp_dim,
-        #         scaling=None,
-        #         right=False,
-        #         stride=1,
-        #         skip=0,
-        #     )
-        #     self.ndim = self.coordinates.dimension()
-
-        elif self.dimReduceMethod == "tica" or self.dimReduceMethod == "vamp":
-
-            # TODO: Streaming implementation, a la
-            #  https://deeptime-ml.github.io/latest/api/generated/deeptime.decomposition.TICA.html#deeptime.decomposition.TICA.partial_fit
-
-            # TODO: Pre-allocate these.. but how can I get their sizes? And, they're ragged
-            trajs_start = []
-            trajs_end = []
-            weights = []
-
-            if last_iter is None:
-                last_iter = self.maxIter
-
-            for iteration in range(first_iter, last_iter, fine_stride):
-
-                # iter_coords = self.get_iter_coordinates(iteration)
-                self.load_iter_data(iteration)
-                self.get_transition_data_lag0()
-
-                coords_from = self.coordPairList[:, :, :, 0]
-                coords_to = self.coordPairList[:, :, :, 1]
-
-                # If  no good coords in this iteration, skip it
-                # if iter_coords.shape[0] == 0:
-                #     continue
-
-                processed_start = self.processCoordinates(coords_from)
-                processed_end = self.processCoordinates(coords_to)
-                trajs_start.extend(processed_start)
-                trajs_end.extend(processed_end)
-                weights.extend(self.weightList)
-
-            weights = np.array(weights)
-
-            if self.dimReduceMethod == "tica":
-                self.coordinates = TICA(
-                    lagtime=1, var_cutoff=variance_cutoff, scaling="kinetic_map"
-                )
-            elif self.dimReduceMethod == "vamp":
-                self.coordinates = VAMP(
-                    lagtime=1, var_cutoff=variance_cutoff, scaling="kinetic_map"
-                )
-
-            # self.coordinates.fit(trajs)
-            log.info(f"Performing weighted {self.dimReduceMethod}")
-            # print(f"Performing weighted TICA with weights {weights.shape} and trajs {trajs.shape}")
-
-            # Weights are not currently supported in VAMP
-            #   See: https://github.com/deeptime-ml/deeptime/blob/main/deeptime/covariance/util/_running_moments.py#L247
-            if not use_weights or self.dimReduceMethod == "vamp":
-                weights = None
-
-            self.coordinates.fit_from_timeseries(
-                (np.array(trajs_start), np.array(trajs_end)), weights=weights
-            )
-
-            # Note: ndim is only used in one place, and it's a deprecated obsolete function
-            self.ndim = self.coordinates.model.output_dimension
-
-            log.info(
-                f"Weighted {self.dimReduceMethod} will reduce "
-                f"{self.coordinates._model._instantaneous_coefficients.shape[0]} to {self.ndim} components."
-            )
-
-        elif self.dimReduceMethod == "none":
-            self.ndim = int(3 * self.nAtoms)
-
-            # TODO: Why is this unused?
-            # data = self.all_coords.reshape(-1, self.ndim)
-            self.coordinates = self.Coordinates()
-            # self.coordinates.transform = self.processCoordinates
-
-        else:
-            raise NotImplementedError
-
-    class Coordinates(object):
-        """
-        Fake Coordinates class, in case you don't want to use either PCA or VAMP for dimensionality reduction
-        """
-
-        # The class "constructor" - It's actually an initializer
-        def __init__(self):
-            self.explanation = "coordinate object"
-
-        def transform(self, coords):
-            return coords
-
-    def reduceCoordinates(self, coords):
-        """
-        This performs the dimensionality reduction.
-
-        dimReduce() defines self.coordinates, which is an object that has a .transform() function that produces the
-        reduced data.
-
-        reduceCoordinates() actually uses that coordinates object, transforms the coordinates,
-        and returns the reduced data.
-
-        The reduced coordinates are then stored in /auxdata for each iteration.
-
-        Parameters
-        ----------
-        coords: array-like
-            Array of coordinates to reduce.
-
-        Returns
-        -------
-        Reduced data
-        """
-
-        # log.debug("Reducing coordinates")
-
-        # TODO: This list should not be stored here, this should be a class attribute or something
-        if (
-            self.dimReduceMethod == "none"
-            or self.dimReduceMethod == "pca"
-            or self.dimReduceMethod == "vamp"
-            or self.dimReduceMethod == "tica"
-        ):
-            coords = self.processCoordinates(coords)
-            coords = self.coordinates.transform(coords)
-            return coords
-
-        raise Exception("dimReduceMethod undefined in reduceCoordinates")
 
     def update_cluster_structures(self):
         """
@@ -2555,1251 +1261,6 @@ class modelWE:
         log.debug("Cluster structure mapping completed.")
         log.debug(f"Cluster keys are {cluster_structures.keys()}")
 
-    def do_clustering(self, arg):
-
-        kmeans_model, iters_to_use, cluster_args, processCoordinates = arg
-
-        # Cast to a list in case it's a np array
-        iters_to_use = list(iters_to_use)
-
-        iteration = iters_to_use[0]
-
-        min_coords = 1
-
-        # The first time we cluster, we need at least n_clusters datapoints.
-        # Before the first clustering has been done, cluster_centers_ is unset, so use that to know if  we're  on the
-        #   first round.
-        if not hasattr(kmeans_model, "cluster_centers_"):
-            log.debug(
-                f"First batch to k-means, need a minimum of {cluster_args['n_clusters']} segments"
-            )
-            min_coords = cluster_args["n_clusters"]
-
-        iter_coords = self.get_iter_coordinates(iteration)
-
-        used_iters = 0
-        # Keep adding coords until you have more than your components
-        while iter_coords.shape[0] <= min_coords:
-
-            iteration = iters_to_use.pop(0)
-
-            used_iters += 1
-
-            _iter_coords = self.get_iter_coordinates(iteration)
-            if _iter_coords.shape[0] == 0:
-                continue
-
-            iter_coords = np.append(iter_coords, _iter_coords, axis=0)
-
-            log.debug(f"Have {iter_coords.shape[0]}, need {min_coords}")
-
-        transformed_coords = self.coordinates.transform(processCoordinates(iter_coords))
-
-        kmeans_model.partial_fit(transformed_coords)
-
-        return kmeans_model, used_iters
-
-    def do_discretization(self, arg):
-
-        kmeans_model, iteration, processCoordinates = arg
-
-        iter_coords = self.get_iter_coordinates(iteration)
-
-        used_iters = 0
-        # Keep adding coords until you have more than your components
-        # This actually isn't necessary here like it is for the clustering, I could just return an empty array
-        while iter_coords.shape[0] < 1:
-
-            used_iters += 1
-
-            _iter_coords = self.get_iter_coordinates(iteration + used_iters)
-            if _iter_coords.shape[0] == 0:
-                continue
-
-            iter_coords = np.append(iter_coords, _iter_coords, axis=0)
-
-        transformed_coords = self.coordinates.transform(processCoordinates(iter_coords))
-
-        dtrajs = kmeans_model.predict(transformed_coords)
-
-        # log.debug(f"Took {used_iters} extra iterations to process iteration {iteration}")
-
-        return dtrajs, used_iters
-
-    @ray.remote
-    def do_ray_discretization(model, kmeans_model, iteration, processCoordinates):
-
-        # model_id, kmeans_model_id, iteration, processCoordinates_id = arg
-
-        # self = ray.get(model_id)
-        # kmeans_model = ray.get(kmeans_model_id)
-        # processCoordinates = ray.get(processCoordinates_id)
-        # self = model
-
-        # Need to do this so the model's transformation array is writable -- otherwise predict chokes
-        #   with 'buffer source array is read-only'.
-        kmeans_model = deepcopy(kmeans_model)
-
-        iter_coords = model.get_iter_coordinates(iteration)
-
-        # If there are no coords for this iteration, return None
-        if iter_coords.shape[0] == 0:
-            return None, 0, iteration
-
-        # Otherwise, apply the k-means model and discretize
-        transformed_coords = model.coordinates.transform(
-            processCoordinates(iter_coords)
-        )
-        dtrajs = kmeans_model.predict(transformed_coords)
-
-        return dtrajs, 1, iteration
-
-    def cluster_coordinates(
-        self,
-        n_clusters,
-        streaming=False,
-        first_cluster_iter=None,
-        use_ray=False,
-        stratified=True,
-        iters_to_use=None,
-        store_validation_model=False,
-        **_cluster_args,
-    ):
-
-        self.clustering_method = None
-
-        log.info(
-            "Be aware: Number of cluster centers is an important parameter, and can drastically affect model "
-            "quality. We recommend examining block-validation results with a range of numbers of clusters, to "
-            "check for overfitting."
-        )
-
-        if stratified:
-            log.info("Beginning stratified clustering.")
-            self.clustering_method = "stratified"
-            self.cluster_stratified(
-                n_clusters=n_clusters,
-                streaming=streaming,
-                first_cluster_iter=first_cluster_iter,
-                use_ray=use_ray,
-                iters_to_use=iters_to_use,
-                **_cluster_args,
-            )
-
-        # Make sure you know what you're doing if using this!
-        else:
-            log.info(
-                "Beginning aggregate clustering. (Beware! This can produce poor clustering)"
-            )
-            self.clustering_method = "aggregated"
-            self.cluster_aggregated(
-                n_clusters=n_clusters,
-                streaming=streaming,
-                first_cluster_iter=first_cluster_iter,
-                use_ray=use_ray,
-                iters_to_use=iters_to_use,
-                **_cluster_args,
-            )
-
-        # Do this so you can use it for building the validation models.
-        #   We directly modify the clusters in cleaning, so this is the easiest way of recreating the original,
-        #   "unclean" validation model.
-        if store_validation_model:
-            self.post_cluster_model = deepcopy(self)
-
-    def cluster_aggregated(
-        self,
-        n_clusters,
-        streaming=False,
-        first_cluster_iter=None,
-        use_ray=False,
-        iters_to_use=None,
-        **_cluster_args,
-    ):
-        """
-        Use k-means to cluster coordinates into `n_clusters` cluster centers, and saves the resulting cluster object
-        to a file.
-
-        Saved cluster file is named
-        "<`self.model_name`>_clusters_s<`self.first_iter`>_e<`self.last_iter`>_nC<`self.n_clusters`>.h5"
-
-        Updates:
-            - `self.n_clusters`
-            - `self.clusters`
-            - `self.clusterFile`
-
-        Parameters
-        ----------
-        n_clusters: int
-            Number of cluster centers to use.
-
-        streaming: boolean
-            Whether to stream k-means clustering, or load all from memory.
-            Not supported for vamp dimensionality reduction.
-
-        **_cluster_args:
-            Keyword arguments that will be passed directly to cluster_kmeans
-
-        Returns
-        -------
-
-        """
-
-        log.debug(f"Doing clustering on {n_clusters} clusters")
-
-        # Have to do this instead of setting a default argument so that we can specify custom iterations to use
-        if first_cluster_iter is None:
-            first_cluster_iter = 1
-
-        self.n_clusters = n_clusters
-        self.first_cluster_iter = first_cluster_iter
-        # streaming = False
-
-        # This is a bit of a gnarly chain of if statements, but it handles checking whether a user explicitly passed
-        #   these arguments.
-        # If you haven't specified a range of iterations to use, then go with first_cluster_iter
-        if iters_to_use is None and first_cluster_iter is None:
-            first_cluster_iter = 1
-            iters_to_use = range(first_cluster_iter, self.maxIter)
-        elif iters_to_use is None and first_cluster_iter is not None:
-            iters_to_use = range(first_cluster_iter, self.maxIter)
-        elif iters_to_use is not None and first_cluster_iter is not None:
-            log.error(
-                "Conflicting parameters -- either iters_to_use OR first_cluster_iter should be provided, not both."
-            )
-        else:
-            # iters_to_use was provided, and first_cluster_iter was not
-            pass
-
-        if "metric" in _cluster_args.keys() or "k" in _cluster_args.keys():
-            log.error(
-                "You're passing pyemma-style arguments to k-means. K-means now uses sklearn, please update"
-                "your code accordingly."
-            )
-            raise Exception(
-                "PyEmma style arguments passed to kmeans, which is now based on sklearn"
-            )
-
-        # Set some default arguments, and overwrite them with the user's choices if provided
-        cluster_args = {
-            "n_clusters": n_clusters,
-            "max_iter": 100,
-        }
-        cluster_args.update(_cluster_args)
-
-        if streaming and not self.dimReduceMethod == "vamp":
-            cluster_model = mini_kmeans(**cluster_args)
-            # TODO: Any mini_batch_kmeans specific arguments?
-
-        elif streaming and self.dimReduceMethod == "vamp":
-            log.warning(
-                "Streaming clustering is not supported for dimReduceMethod 'none'. Using standard k-means."
-            )
-            cluster_model = kmeans(**cluster_args)
-            streaming = False
-
-        else:
-            cluster_model = kmeans(**cluster_args)
-
-        if self.dimReduceMethod == "none":
-
-            if not streaming:
-
-                # TODO: Fix this to correctly use processCoordinates, like in the streaming case
-                raise NotImplementedError(
-                    "Non-streaming dimreduce None is currently broken"
-                )
-
-                _data = [
-                    self.get_iter_coordinates(iteration).reshape(
-                        -1, self.coord_ndim * self.nAtoms
-                    )
-                    for iteration in iters_to_use
-                ]
-                stacked_data = np.vstack(_data)
-
-                if self.nAtoms > 1:
-                    self.clusters = cluster_model.fit(stacked_data)
-
-                # Else here is a little sketchy, but fractional nAtoms is useful for some debugging hacks.
-                else:
-                    self.clusters = cluster_model.fit(stacked_data)
-
-                self.dtrajs = [
-                    cluster_model.predict(iter_trajs) for iter_trajs in _data
-                ]
-
-            else:
-                self.dtrajs = []
-
-                extra_iters_used = 0
-                for iter_idx, iteration in enumerate(
-                    tqdm.tqdm(iters_to_use, desc="Clustering")
-                ):
-
-                    if extra_iters_used > 0:
-                        extra_iters_used -= 1
-                        log.debug(f"Already processed  iter  {iteration}")
-                        continue
-
-                    with concurrent.futures.ProcessPoolExecutor(
-                        max_workers=1, mp_context=mp.get_context("fork")
-                    ) as executor:
-                        cluster_model, extra_iters_used = executor.submit(
-                            self.do_clustering,
-                            [
-                                cluster_model,
-                                [iters_to_use[iter_idx]],
-                                cluster_args,
-                                self.processCoordinates,
-                            ],
-                        ).result()
-
-                self.clusters = cluster_model
-
-                # Now compute dtrajs from the final model
-                # TODO: Do I want to always discretize the full data?
-                for iteration in range(1, self.maxIter):
-                    iter_coords = self.get_iter_coordinates(iteration)
-                    transformed_coords = self.coordinates.transform(
-                        self.processCoordinates(iter_coords)
-                    )
-
-                    self.dtrajs.append(cluster_model.predict(transformed_coords))
-
-        elif self.dimReduceMethod == "pca" and streaming:
-
-            self.dtrajs = []
-
-            # continued = False
-            extra_iters_used = 0
-            for iter_idx, iteration in enumerate(
-                tqdm.tqdm(iters_to_use, desc="Clustering")
-            ):
-
-                if extra_iters_used > 0:
-                    extra_iters_used -= 1
-                    log.debug(f"Already processed  iter  {iteration}")
-                    continue
-
-                with concurrent.futures.ProcessPoolExecutor(
-                    max_workers=1, mp_context=mp.get_context("fork")
-                ) as executor:
-                    cluster_model, extra_iters_used = executor.submit(
-                        self.do_clustering,
-                        [
-                            cluster_model,
-                            iters_to_use[iter_idx:],
-                            cluster_args,
-                            self.processCoordinates,
-                        ],
-                    ).result()
-
-            self.clusters = cluster_model
-
-            # Now compute dtrajs from the final model
-            extra_iters_used = 0
-
-            # If we're not using Ray, then calculate serially
-            if not use_ray:
-                for iteration in tqdm.tqdm(
-                    range(1, self.maxIter), desc="Discretization"
-                ):
-
-                    if extra_iters_used > 0:
-                        extra_iters_used -= 1
-                        log.debug(f"Already processed  iter  {iteration}")
-                        continue
-
-                    with concurrent.futures.ProcessPoolExecutor(
-                        max_workers=1, mp_context=mp.get_context("fork")
-                    ) as executor:
-                        dtrajs, extra_iters_used = executor.submit(
-                            self.do_discretization,
-                            [cluster_model, iteration, self.processCoordinates],
-                        ).result()
-
-                    self.dtrajs.append(dtrajs)
-
-            # If we want to use Ray
-            else:
-
-                # First, connect to the ray cluster
-                self.check_connect_ray()
-
-                # Submit all the discretization tasks to the cluster
-                task_ids = []
-
-                model_id = ray.put(self)
-                cluster_model_id = ray.put(cluster_model)
-                process_coordinates_id = ray.put(self.processCoordinates)
-
-                # max_inflight = 50
-                for iteration in tqdm.tqdm(
-                    range(1, self.maxIter), desc="Submitting discretization tasks"
-                ):
-
-                    # if len(task_ids) > max_inflight:
-                    #
-                    #     # The number that need to be ready before we can submit more
-                    #     num_ready = iteration - max_inflight
-                    #     ray.wait(task_ids, num_returns=num_ready)
-
-                    _id = self.do_ray_discretization.remote(
-                        model_id, cluster_model_id, iteration, process_coordinates_id
-                    )
-                    task_ids.append(_id)
-
-                # As they're completed, add them to dtrajs
-                dtrajs = [None] * (self.maxIter - 1)
-
-                # Do these in bigger batches, dtrajs aren't very big
-
-                with tqdm.tqdm(
-                    total=len(task_ids), desc="Retrieving discretized trajectories"
-                ) as pbar:
-                    while task_ids:
-                        result_batch_size = 50
-                        result_batch_size = min(result_batch_size, len(task_ids))
-
-                        # Returns the first ObjectRef that is ready, with a 20s timeout
-                        finished, task_ids = ray.wait(
-                            task_ids, num_returns=result_batch_size, timeout=20
-                        )
-                        results = ray.get(finished)
-
-                        for dtraj, _, iteration in results:
-                            dtrajs[iteration - 1] = dtraj
-                            pbar.update(1)
-                            pbar.refresh()
-
-                # Remove all empty elements from dtrajs and assign to self.dtrajs
-                self.dtrajs = [dtraj for dtraj in dtrajs if dtraj is not None]
-
-        elif self.dimReduceMethod == "vamp" and streaming:
-
-            raise NotImplementedError("VAMP + streaming clustering is not supported.")
-
-        elif (
-            self.dimReduceMethod == "pca"
-            or self.dimReduceMethod == "vamp"
-            and not streaming
-        ):
-
-            if self.dimReduceMethod == "pca":
-                transformed_data = []
-
-                for iteration in iters_to_use:
-                    iter_coords = self.get_iter_coordinates(iteration)
-
-                    # Skip if  this is an empty iteration
-                    if iter_coords.shape[0] == 0:
-                        continue
-
-                    transformed_data.append(
-                        self.coordinates.transform(self.processCoordinates(iter_coords))
-                    )
-                    log.debug(f"Appended trajs for iter {iteration}")
-
-                stacked_data = np.vstack(transformed_data)
-                self.clusters = cluster_model.fit(stacked_data)
-                self.dtrajs = [
-                    cluster_model.predict(iter_trajs) for iter_trajs in transformed_data
-                ]
-
-            # TODO: Make sure this is returned in the shape (iteration, ...)
-            elif self.dimReduceMethod == "vamp":
-                log.warning("Clustering VAMP-reduced data still very experimental!")
-
-                transformed_data = self.coordinates.get_output()
-
-                self.clusters = cluster_model.fit(transformed_data)
-                self.dtrajs = [
-                    cluster_model.predict(iter_trajs) for iter_trajs in transformed_data
-                ]
-
-        self.clusterFile = (
-            self.modelName
-            + "_clusters_s"
-            + str(self.first_iter)
-            + "_e"
-            + str(self.last_iter)
-            + "_nC"
-            + str(self.n_clusters)
-            + ".h5"
-        )
-
-        # self.dtrajs = self.clusters.dtrajs
-        assert self.dtrajs is not None
-
-        log.debug("Clustering completed.")
-        # log.debug(f"Dtrajs: {self.clusters.dtrajs}")
-        # self.clusters.save(self.clusterFile, save_streaming_chain=True, overwrite=True)
-
-    def cluster_stratified(
-        self,
-        n_clusters,
-        streaming=True,
-        first_cluster_iter=None,
-        use_ray=True,
-        bin_iteration=2,
-        iters_to_use=None,
-        user_bin_mapper=None,
-        **_cluster_args,
-    ):
-        """
-        Perform full-stratified clustering, enforcing independent clustering for trajectories within each WE bin.
-
-        Parameters
-        ----------
-        n_clusters: int
-            Number of cluster centers per WE bin.
-
-        streaming: bool (default True)
-            Must be True for now.
-
-        first_cluster_iter: int
-            Iteration to start clustering on.
-            Discretization is always done on all iteration, but it may not be necessary or desirable to use all to
-            cluster.
-
-        use_ray: bool (default True)
-            Must be True for now.
-
-        bin_iteration: int (default 2)
-            Iteration to obtain bin definitions from.
-
-        iters_to_use: array-like (default [1, model.maxIter])
-            Specific set of iterations to cluster on.
-
-        **_cluster_args:
-            Arguments passed through to sklearn.cluster.MiniBatchKMeans
-
-        Returns
-        -------
-
-        TODO
-        ----
-        Allow directly passing in a bin mapper or equivalent here. Way easier than trying to reverse engineer one.
-        Require it if the existing is not supported.
-        """
-
-        # This provides backwards compatibility when loading WESTPA data from simulations where binning lived in westpa.binning
-        # It's a little sketchy, but it import westpa.tools.binning under westpa.binning
-        import importlib
-        import westpa
-
-        westpa.binning = importlib.import_module("westpa.tools.binning", "westpa.tools")
-
-        if user_bin_mapper is not None:
-            log.info("Loading user-specified bin mapper for stratified clustering.")
-            bin_mapper = user_bin_mapper
-        else:
-            log.info(
-                f"Obtaining bin definitions from iteration {bin_iteration} in file {self.fileList[0]}"
-            )
-            iteration = analysis.Run(self.fileList[0]).iteration(bin_iteration)
-            log.info("Loading pickled bin mapper from H5 for stratified clustering...")
-            bin_mapper = iteration.bin_mapper
-
-            # Problem: I need a consistent set of bins, and some bin mappers may not return that! Some may re-calculate bins
-            #    on-the-fly whenever data is passed to them, which will produce different bins for each iteration.
-            #   In particular, ones that may are really any functional mapper:
-            #   -   MABBinMapper will by construction never return bins that have everything populated, and by default,
-            #       calculates boundaries on the fly
-            #   -   PiecewiseBinMapper/VectorizingFuncBinMapper/FuncBinMapper are also function-based so they may do whatever
-            #   Ones that won't:
-            #   -   VoronoiBinMapper takes in a set of centers when it's initialized, so it'll be consistent
-            #   -   Rectilinear is of course always consistent
-            # Ideally, the best approach here would be to go through your bin mapper, recursing as necessary, and replacing
-            #   any of the above functional bin mapper types with static bin mappers.
-
-            if type(bin_mapper) not in SUPPORTED_MAPPERS:
-                log.warning(
-                    f"{type(bin_mapper)} mapper loaded, but supported mappers are {SUPPORTED_MAPPERS} and others may"
-                    f"produce inconsistent bins between iterations. Please provide a supported user_bin_mapper."
-                )
-                raise Exception
-
-            # TODO: Before moving on, make sure it's actually possible to populate each bin with enough segments to
-            #       cluster at least once from the given set of iterations.
-            #   In other words, just load up pcoords iteration by iteration, and count the total number in each bin.
-            #   Note that it's possible you'll end up clustering in a bin once, but then pulling too few in that bin for
-            #       the rest of the iterations after that, or something like that.
-            #   The alternative to this would be in do_stratified_clustering(), if a bin isn't populated, check every
-            #       subsequent iteration and see if it's possible to populate it from them.
-            #   This is a little more flexible/dynamic, however, I think it may potentially require much more looping
-            #       through iterations.
-
-        # Clustering will not be performed in these bins
-        ignored_bins = []
-
-        if not streaming or not use_ray:
-            log.error(
-                "Stratified clustering currently MUST run in streaming mode (and with ray enabled for "
-                "discretization). Enabling both and continuing..."
-            )
-            streaming = True
-            use_ray = True
-
-        # Do the streaming clustering, but cluster each bin individually.
-        stratified_clusters = StratifiedClusters(
-            bin_mapper, self, n_clusters, ignored_bins, **_cluster_args
-        )
-
-        # This is a bit of a gnarly chain of if statements, but it handles checking whether a user explicitly passed
-        #   these arguments.
-        # If you haven't specified a range of iterations to use, then go with first_cluster_iter
-        if iters_to_use is None and first_cluster_iter is None:
-            first_cluster_iter = 1
-            iters_to_use = range(first_cluster_iter, self.maxIter)
-        elif iters_to_use is None and first_cluster_iter is not None:
-            iters_to_use = range(first_cluster_iter, self.maxIter)
-        elif iters_to_use is not None and first_cluster_iter is not None:
-            log.error(
-                "Conflicting parameters -- either iters_to_use OR first_cluster_iter should be provided, not both."
-            )
-        else:
-            # iters_to_use was provided, and first_cluster_iter was not
-            pass
-
-        # ## Build the clustering model
-        self.dtrajs = []
-        extra_iters_used = 0
-        all_filled_bins = set()
-        all_unfilled_bins = set()
-        for iter_idx, iteration in enumerate(
-            tqdm.tqdm(iters_to_use, desc="Clustering")
-        ):
-
-            if extra_iters_used > 0:
-                extra_iters_used -= 1
-                log.debug(f"Already processed  iter  {iteration}")
-                continue
-
-            ignored_bins = []
-
-            with concurrent.futures.ProcessPoolExecutor(
-                max_workers=1, mp_context=mp.get_context("fork")
-            ) as executor:
-
-                try:
-                    (
-                        stratified_clusters,
-                        extra_iters_used,
-                        filled_bins,
-                        unfilled_bins,
-                    ) = executor.submit(
-                        self.do_stratified_clustering,
-                        [
-                            self,
-                            stratified_clusters,
-                            iters_to_use[iter_idx:],
-                            self.processCoordinates,
-                            ignored_bins,
-                        ],
-                    ).result()
-
-                except AssertionError as e:
-                    # If we succeeded in passing this loop at least once, then all our bins have *something* in them.
-                    # TODO: The better way to handle this is as long as you've clustered something, you can do piecemeal
-                    #   (i.e. don't have to populate every bin after the first time)
-                    if iter_idx == 0:
-                        log.info(
-                            f"Failed with {iter_idx} + {extra_iters_used} vs len {(len(iters_to_use))}"
-                        )
-                        raise e
-                    else:
-                        log.info(
-                            "Clustering couldn't use last iteration, not all bins filled."
-                        )
-
-                all_filled_bins.update(filled_bins)
-                all_unfilled_bins.update(unfilled_bins)
-
-        # all_filled_bins holds every bin that was clustered in
-        # all_unfilled_bins holds any bin that was ever attempted, but unfilled
-        # so, get the unfilled that were unfilled always, and never got clustered in
-        true_unfilled = np.setdiff1d(range(bin_mapper.nbins), list(all_filled_bins))
-        log.debug(f"Filled bins are {all_filled_bins}")
-        log.debug(f"Unfilled bins were {all_unfilled_bins}")
-        log.debug(f"True unfilled bins are {true_unfilled}")
-
-        for unfilled_bin_idx in true_unfilled:
-            remap_bin = self.find_nearest_bin(
-                bin_mapper, unfilled_bin_idx, list(all_filled_bins)
-            )
-            stratified_clusters.we_remap[unfilled_bin_idx] = remap_bin
-            log.debug(f"Remapped {unfilled_bin_idx} to {remap_bin}")
-
-        # make sure this doesn't mess up we_remap later.. I think it should be fine, when I write to we_remap in the cleaning
-        #       it should identify a superset of the bins identified here.
-
-        self.clusters = stratified_clusters
-        self.clusters.model = self
-
-        ## The below -1 was removed, because now I may specify a custom set of WE bins
-        ## -1 because we don't cluster in the target
-
-        self.n_clusters = n_clusters * (bin_mapper.nbins)
-
-        self.clusters.toggle = False
-
-        self.launch_ray_discretization()
-
-    @staticmethod
-    def find_nearest_bin(bin_mapper, bin_idx, filled_bins):
-        """
-        Given a bin mapper, find the bin closest to bin_idx (that isn't bin_idx).
-
-        Do this Voronoi-style by obtaining a set of bin centers, and finding which center bin_idx is closest to.
-
-        Parameters
-        ----------
-        bin_mapper
-        bin_idx
-
-        Returns
-        -------
-        Index of the closest bin.
-
-        TODO
-        ----
-        Note in the documentation that this can be overriden for finer control over empty bin mapping, if so desired.
-        """
-
-        assert len(filled_bins) > 0, (
-            "Can't find nearest populated bin -- no WE bins are populated with clusters! "
-            "Try fewer clusters/bin."
-        )
-
-        assert type(bin_mapper) in [
-            VoronoiBinMapper,
-            RectilinearBinMapper,
-        ], f"{type(bin_mapper)} is unsupported!"
-
-        if type(bin_mapper) is VoronoiBinMapper:
-
-            centers = bin_mapper.centers
-            distance_function = bin_mapper.dfunc
-
-        elif type(bin_mapper) is RectilinearBinMapper:
-
-            def _rmsd(point, _centers):
-                return np.sqrt(np.mean(np.power(point - _centers, 2), axis=1))
-
-            distance_function = _rmsd
-
-            bounds = np.array(bin_mapper.boundaries)
-            _centers = []
-            for dim in bounds:
-                _centers.append(dim[:-1] + (dim[1:] - dim[:-1]) / 2)
-            centers = (
-                np.array(np.meshgrid(*_centers)).T.squeeze().reshape(-1, len(bounds))
-            )
-
-        # Remove both the bin you're looking at, and any other unfilled bins
-        # all_ignored = np.concatenate([[bin_idx], unfilled_bins])
-
-        # Ignore any bin that isn't an explicitly provided filled bin
-        all_ignored = np.setdiff1d(range(centers.shape[0]), filled_bins)
-        other_centers = np.delete(centers, all_ignored, axis=0)
-
-        closest = np.argmin(distance_function(centers[bin_idx], other_centers))
-
-        # Increment index if it's past the ones we deleted
-        for _bin_idx in sorted(all_ignored):
-            if closest >= _bin_idx:
-                closest += 1
-
-        return closest
-
-    def do_stratified_clustering(self, arg):
-        """
-        Perform the full-stratified clustering.
-
-        This works as follows:
-
-            1. Pull coordinates from the first iteration to process
-            2. Assign each segment to a WE bin, using its pcoord and the bin_mapper associated with this StratifiedCluster
-                object.
-            3a. If any of the seen WE bins have fewer segments than cluster centers, and there are more iterations left
-                to process, repeat from 1.
-                - Note that this may add NEW seen bins, and those new seen bins may not be full yet -- so more iterations
-                    may be required.
-            3b. If any seen WE bins have fewer segments than cluster centers, **but no iterations are left to process**,
-                then assign each structure in an "unfilled" WE bin to the "filled" WE bin with the closest index.,
-
-            At this point, we have a set of structures and the WE bins they're associated with, and each WE bin has a
-            number of structures equal to or greater than the target number of cluster centers.
-
-            Within each WE bin:
-
-            4. Apply dimensionality reduction to structures
-            5. Update clustering for that bin using this set of dimensionality reduced coordinates.
-        """
-
-        self, kmeans_models, iters_to_use, processCoordinates, ignored_bins = arg
-
-        # Cast to a list in case it's a np array
-        iters_to_use = list(iters_to_use)
-
-        bin_mapper = kmeans_models.bin_mapper
-
-        min_coords = kmeans_models.cluster_args["n_clusters"]
-
-        # The number of populated bins is the number of total bins - 1
-        all_bins_have_segments = False
-
-        # Until all bins are populated
-        used_iters = -1
-        iter_coords = []
-        unique_bins = np.array([])
-        counts = np.array([])
-        we_bin_assignments = []
-
-        # Maybe not even necessary to track iter_coords, just _iter_coords.
-        # The problem with it as it stands is that assert -- imagine if I have to grab a second iteration
-        while not all_bins_have_segments:
-            unfilled_bins = []
-
-            # This may cover the same use case as iteration > self.maxIter below
-            try:
-                iteration = iters_to_use.pop(0)
-            except IndexError:
-                log.warning(
-                    f"At iteration {iteration} (pulled {used_iters} extra), couldn't get segments in all bins, and no "
-                    f"iterations left."
-                )
-
-                # Which bin didn't have enough clusters?
-                unfilled_bins = unique_bins[counts < min_coords]
-                filled_bins = np.setdiff1d(unique_bins, unfilled_bins)
-
-                # Find the nearest non-empty bin
-                for unfilled_bin in unfilled_bins:
-
-                    nearest_filled_bin = self.find_nearest_bin(
-                        bin_mapper, unfilled_bin, list(filled_bins)
-                    )
-
-                    unfilled_bin_indices = np.where(we_bin_assignments == unfilled_bin)
-                    log.warning(
-                        f"Remapping {len(unfilled_bin_indices)} segments from unfilled bin {unfilled_bin} to "
-                        f"{nearest_filled_bin} for stratified clustering"
-                    )
-                    we_bin_assignments[unfilled_bin_indices] = nearest_filled_bin
-
-                # Remove unfilled bins from unique_bins
-                unique_bins = filled_bins
-
-                break
-
-            if used_iters > -1:
-                log.debug(
-                    f"Still missing segs in bin {unique_bins[counts < min_coords]}. Pulled {used_iters+1} extra iter"
-                )
-                log.debug(f"Had {list(zip(unique_bins, counts))}, need {min_coords}")
-
-            if iteration > self.maxIter:
-                # TODO: Is this always a deal-breaker?
-                log.warning(
-                    f"At iteration {iteration} (pulled {used_iters} extra), couldn't get segments in all bins, and no "
-                    f"iterations left"
-                )
-                break
-
-            used_iters += 1
-            _iter_coords = self.get_iter_coordinates(iteration)
-            _seg_weights = self.seg_weights[iteration]
-
-            if used_iters == 0:
-                iter_coords = _iter_coords
-                seg_weights = _seg_weights
-                pcoords = [x for x in self.pcoord0List]
-            else:
-                iter_coords = np.append(iter_coords, _iter_coords, axis=0)
-                pcoords.extend(self.pcoord0List)
-                if self.use_weights_in_clustering:
-                    seg_weights = np.append(seg_weights, _seg_weights, axis=0)
-                log.debug(
-                    f"After extension, pcoords: {len(pcoords)}, iter_coords: {iter_coords.shape}. Ignored: {ignored_bins}. Mapper: {bin_mapper}"
-                )
-
-            # Map coords to WE bins
-            pcoord_array = np.array(pcoords)
-            # seg_weight_array = np.array(seg_weights)
-            assert (
-                pcoord_array.shape[0] == iter_coords.shape[0]
-            ), f"{pcoord_array.shape}, {iter_coords.shape}"
-
-            if self.use_weights_in_clustering:
-                assert (
-                    seg_weights.shape[0] == seg_weights.shape[0]
-                ), f"{seg_weights.shape}, {iter_coords.shape}"
-
-            # Ignore any segments that are in the basis or target
-            pcoord_is_target = self.is_WE_target(pcoord_array)
-            pcoord_is_basis = self.is_WE_basis(pcoord_array)
-            pcoord_array = pcoord_array[~(pcoord_is_target | pcoord_is_basis)]
-
-            if len(pcoord_array) > 0:
-                we_bin_assignments = bin_mapper.assign(pcoord_array)
-            else:
-                log.debug(
-                    "No coordinates outside of basis/target, we_bin_assignments will be empty and clustering"
-                    f" will be skipped for this iteration. ({sum(pcoord_is_target)} in target, {sum(pcoord_is_basis)} in basis)"
-                )
-                we_bin_assignments = np.array([])
-
-            unique_bins, counts = np.unique(we_bin_assignments, return_counts=True)
-            all_bins_have_segments = np.all(counts >= min_coords)
-
-        # By now, I have some segments in each WE bin.
-        # Now, do clustering within each WE bin, and only cluster in WE bins that things were actually assigned to
-        for i, _bin in enumerate(unique_bins):
-
-            segs_in_bin = np.argwhere(we_bin_assignments == _bin)
-
-            transformed_coords = self.coordinates.transform(
-                processCoordinates(np.squeeze(iter_coords[segs_in_bin]))
-            )
-
-            if self.use_weights_in_clustering:
-                weights = seg_weights[segs_in_bin].squeeze()
-            else:
-                weights = None
-
-            if weights is not None:
-                log.debug(weights.shape)
-            else:
-                log.debug("Weights are None")
-
-            try:
-                kmeans_models.cluster_models[_bin].partial_fit(
-                    transformed_coords, sample_weight=weights
-                )
-
-            except ValueError as e:
-                log.info(f"Was on bin {_bin}")
-                log.error(f"Error fitting k-means to bin {_bin}")
-                raise e
-
-        return kmeans_models, used_iters, unique_bins, unfilled_bins
-
-    def organize_stratified(self, use_ray=True):
-        """
-        Alternative to organize_fluxMatrix, for stratified clustering.
-
-        Parameters
-        ----------
-        use_ray: bool
-            If True, use Ray when calculating the cleaned fluxmatrix.
-
-        Returns
-        -------
-
-        """
-
-        # Find disconnected states
-        fmatrix_original = self.fluxMatrixRaw.copy()
-
-        fmatrix = self.fluxMatrixRaw.copy()
-        fmatrix[-1, -2] = 1.0
-        connected_sets = find_connected_sets(fmatrix, directed=True)
-
-        start_cleaning_idx = 1
-
-        if len(connected_sets) == 1:
-            log.info("Nothing to clean")
-            #         return
-            states_to_remove = []
-        else:
-
-            # # Experimental
-            # # I've seen cases where I have two large disconnected sets of similar size, with problems that arise from cleaning
-            # #   one entirely. I think maybe this is like two sides of a barrier, with weak connection between, so when you
-            # #   do the cleaning basically everything on one side gets lumped together.
-            # # To try and ameliorate this, what if we group together large clusters, and bank on the fact that after we clean
-            # #   all the small isolated clusters, those large ones will be connected?
-            # # This doesn't work :( After cleaning, these sets remain disconnected
-            #
-            # set_sizes = np.array([len(_set) for _set in connected_sets])
-            # primary_set_size = set_sizes[0]
-            #
-            # # Keep any sets that are at least half the size of the primary set
-            # start_cleaning_idx = np.argmin(set_sizes > (primary_set_size//2)).flatten()[0]
-            # log.info(f"Merging sets of size {set_sizes[:start_cleaning_idx]} and cleaning the rest")
-            # #
-
-            log.debug(
-                f"Cleaning states {np.concatenate(connected_sets[start_cleaning_idx:])}"
-            )
-            states_to_remove = np.concatenate(connected_sets[start_cleaning_idx:])
-
-        # Does NOT include the basis/target -- none of these numbers of cluster centers should,
-        #   since they're explicitly not clustered in
-        pre_cleaning_n_clusters_per_bin = [
-            len(cluster_model.cluster_centers_)
-            if hasattr(cluster_model, "cluster_centers_")
-            else 0
-            for cluster_model in self.clusters.cluster_models
-        ]
-
-        log.debug(
-            f"Pre-cleaning clusters per bin: "
-            f"{list(zip(range(len(pre_cleaning_n_clusters_per_bin)), pre_cleaning_n_clusters_per_bin))}"
-        )
-
-        empty_we_bins = set()
-        # target, basis = self.clusters.target_bins, self.clusters.basis_bins
-        # Go through each WE bin, finding which clusters within it are not in the connected set
-        #    and removing them.
-        for we_bin in range(self.clusters.bin_mapper.nbins):
-
-            consecutive_index = we_bin
-
-            offset = sum(pre_cleaning_n_clusters_per_bin[:we_bin])
-
-            # Get all the clusters that would be contained in this bin
-            # clusters_in_bin = range(offset, offset + self.clusters.n_clusters_per_bin)
-            n_clusters_in_bin = 0
-            if hasattr(
-                self.clusters.cluster_models[consecutive_index], "cluster_centers_"
-            ):
-                n_clusters_in_bin = len(
-                    self.clusters.cluster_models[consecutive_index].cluster_centers_
-                )
-            # Otherwise, this is an uninitialized bin.
-            # Bins can be uninitialized in 2 cases -- completely unvisited, or a basis/target
-            # Segments in a basis/target should not be mapped to other bins -- however, this is short-circuited in
-            #   StratifiedClusters.predict(), so even if they're mapped to those cluster centers, it's fine.
-
-            clusters_in_bin = range(
-                offset,
-                offset + n_clusters_in_bin,
-            )
-            log.debug(f"Cluster models len: {len(self.clusters.cluster_models)}")
-            log.debug(
-                f"WE Bin {we_bin} (consec. index {consecutive_index}) contains {n_clusters_in_bin} clusters {clusters_in_bin}"
-            )
-
-            # Find which of the removed clusters are in this
-            bin_clusters_to_clean = np.intersect1d(states_to_remove, clusters_in_bin)
-
-            # Before actually cleaning, check if we're not cleaning anything, or if we're cleaning everything
-
-            # If not cleaning anything, just move on
-            if len(bin_clusters_to_clean) == 0:
-                log.debug(f"not cleaning any clusters from bin {we_bin}")
-
-                if len(clusters_in_bin) == 0:
-                    empty_we_bins.add(we_bin)
-
-                continue
-
-            # If cleaning EVERYTHING, handle this bin differently
-            # We'll just re-map it to a "good" adjacent WE bin
-            elif len(bin_clusters_to_clean) == len(clusters_in_bin):
-                empty_we_bins.add(we_bin)
-
-            else:
-                log.debug(
-                    f"Cleaning {len(bin_clusters_to_clean)} clusters {bin_clusters_to_clean} from WE bin {we_bin}."
-                )
-
-            self.clusters.cluster_models[we_bin].cluster_centers_ = np.delete(
-                self.clusters.cluster_models[we_bin].cluster_centers_,
-                bin_clusters_to_clean - offset,
-                0,
-            )
-
-        log.info(
-            f"Started with {self.n_clusters} clusters, and removed {len(states_to_remove)}"
-        )
-        self.n_clusters = self.n_clusters - len(states_to_remove)
-        log.debug(f"n_clusters is now {self.n_clusters}")
-
-        assert (
-            self.n_clusters > 1
-        ), "All clusters would be cleaned! You probably need more data, fewer clusters, or both."
-
-        # If a WE bin was completely emptied of cluster centers, map it to the nearest non-empty bin
-        populated_we_bins = np.setdiff1d(
-            range(self.clusters.bin_mapper.nbins), list(empty_we_bins)
-        )
-
-        if len(empty_we_bins) > 0:
-            log.warning(f"All clusters were cleaned from bins {empty_we_bins}")
-
-        for empty_we_bin in empty_we_bins:
-
-            # Find the nearest non-empty bin
-            nearest_populated_bin = self.find_nearest_bin(
-                self.clusters.bin_mapper, empty_we_bin, populated_we_bins
-            )
-
-            # Replace self.clusters.cluster_models[empty_we_bin].cluster_centers_ with
-            #   self.clusters.cluster_models[nearest_nonempty_we_bin].cluster_centers_
-            self.clusters.we_remap[empty_we_bin] = nearest_populated_bin
-
-        _running_total = 0
-        for we_bin in range(self.clusters.bin_mapper.nbins):
-
-            try:
-                clusters_in_bin = len(
-                    self.clusters.cluster_models[
-                        self.clusters.we_remap[we_bin]
-                    ].cluster_centers_
-                )
-            except AttributeError as e:
-                log.error(
-                    f"Error obtaining clusters for WE bin {we_bin}, remapped to {self.clusters.we_remap[we_bin]}. "
-                    f"Target {self.clusters.target_bins}, basis {self.clusters.basis_bins}"
-                )
-                raise e
-
-            _running_total += clusters_in_bin
-            log.debug(
-                f"{clusters_in_bin} in bin {we_bin}. Running total: {_running_total}"
-            )
-
-        # Now re-discretize
-        self.clusters.toggle = False
-        self.clusters.processing_from = False
-        self.launch_ray_discretization()
-
-        pcoord_sort_indices = self.get_cluster_centers()
-
-        # And recalculate the flux matrix
-        self.clusters.toggle = True
-        self.clusters.processing_from = True
-        self.get_fluxMatrix(*self._fluxMatrixParams, use_ray=use_ray)
-        self.clusters.processing_from = False
-        self.clusters.toggle = False
-
-        # Set the cleaned matrix as fluxMatrix
-        fluxMatrix = self.fluxMatrixRaw.copy()
-        fluxMatrix = fluxMatrix[pcoord_sort_indices, :]
-        fluxMatrix = fluxMatrix[:, pcoord_sort_indices]
-        self.fluxMatrix = fluxMatrix / np.sum(fluxMatrix)
-
-        self.fluxMatrixRaw = fmatrix_original
-
-        self.indBasis = np.array([self.n_clusters])
-        self.indTargets = np.array([self.n_clusters + 1])
-        self.nBins = self.n_clusters + 2
-
-        self.cluster_mapping = {x: x for x in range(self.n_clusters + 2)}
-
-        # Sanity check that cleaning worked
-        fmatrix = self.fluxMatrix.copy()
-        fmatrix[-1, -2] = 1.0
-        connected_sets = find_connected_sets(fmatrix, directed=True)
-        log.debug(
-            f"After cleaning, shape is {fmatrix.shape} and disconnected sets are: {connected_sets[1:]}"
-        )
-        assert (
-            len(connected_sets[start_cleaning_idx:]) == 0
-        ), "Still not clean after cleaning!"
-
-    def launch_ray_discretization(self):
-        """
-        Apply discretization in parallel, through Ray
-
-        Returns
-        -------
-
-        Todo
-        ====
-        Replace manual discretization calls in cluster_aggregated with this
-        Add flag to toggle between stratified and regular do_ray_discretization
-        """
-
-        self.check_connect_ray()
-
-        self.dtrajs = []
-
-        # Submit all the discretization tasks to the cluster
-        task_ids = []
-
-        if self.pre_discretization_model is None:
-            self.pre_discretization_model = deepcopy(self)
-        else:
-            log.info("Using cached model for discretization")
-
-        model_id = ray.put(self.pre_discretization_model)
-
-        clusters = deepcopy(self.clusters)
-        # It's set inside do_stratified_ray_discretization, though I could do it in either place.
-        clusters.model = None  # self.pre_discretization_model
-        cluster_model_id = ray.put(clusters)
-
-        process_coordinates_id = ray.put(self.processCoordinates)
-
-        # max_inflight = 50
-        for iteration in tqdm.tqdm(
-            range(1, self.maxIter), desc="Submitting discretization tasks"
-        ):
-
-            _id = self.do_stratified_ray_discretization.remote(
-                model_id,
-                cluster_model_id,
-                iteration,
-                process_coordinates_id
-                # self, self.clusters, iteration, self.processCoordinates
-            )
-            task_ids.append(_id)
-
-        # As they're completed, add them to dtrajs
-        dtrajs = [None] * (self.maxIter - 1)
-        pair_dtrajs = [None, None] * (self.maxIter - 1)
-
-        # Do these in bigger batches, dtrajs aren't very big
-
-        with tqdm.tqdm(
-            total=len(task_ids), desc="Retrieving discretized trajectories"
-        ) as pbar:
-            while task_ids:
-                result_batch_size = 50
-                result_batch_size = min(result_batch_size, len(task_ids))
-
-                # Returns the first ObjectRef that is ready, with a 20s timeout
-                finished, task_ids = ray.wait(
-                    task_ids, num_returns=result_batch_size, timeout=20
-                )
-                results = ray.get(finished)
-
-                for (
-                    (parent_dtraj, child_dtraj),
-                    _,
-                    iteration,
-                    target_bins,
-                    basis_bins,
-                ) in results:
-
-                    self.clusters.target_bins.update(target_bins)
-                    self.clusters.basis_bins.update(basis_bins)
-
-                    dtrajs[iteration - 1] = child_dtraj
-
-                    pair_dtrajs[iteration - 1] = list(zip(parent_dtraj, child_dtraj))
-
-                    pbar.update(1)
-                    pbar.refresh()
-
-                del results
-                del finished
-        del model_id
-        del cluster_model_id
-
-        # Remove all empty elements from dtrajs and assign to self.dtrajs
-        self.dtrajs = [dtraj for dtraj in dtrajs if dtraj is not None]
-
-        self.pair_dtrajs = [dtraj for dtraj in pair_dtrajs if dtraj is not None]
-
-        log.info("Discretization complete")
-
     def get_cluster_centers(self):
         """
         Standalone method to obtain average pcoords of all segments in each cluster.
@@ -3872,1931 +1333,6 @@ class modelWE:
         self.targetRMSD_all = np.array(cluster_pcoord_all)[pcoord_sort_indices]
 
         return pcoord_sort_indices
-
-    @ray.remote
-    def do_stratified_ray_discretization(
-        model, kmeans_model, iteration, processCoordinates
-    ):
-
-        # model_id, kmeans_model_id, iteration, processCoordinates_id = arg
-
-        # import sys
-        # import westpa.core.binning
-
-        # sys.modules["westpa.binning"] = sys.modules["westpa.core.binning"]
-        # This is silly -- I need to import westpa.core.binning so it's loaded into sys.modules but the linter
-        #   complains that it's unused... so, use it.
-        # log.debug(f"Loaded {westpa.core.binning}")
-
-        # self = ray.get(model_id)
-        # kmeans_model = ray.get(kmeans_model_id)
-        # processCoordinates = ray.get(processCoordinates_id)
-        self = model
-
-        # Need to do this so the model's transformation array is writable -- otherwise predict chokes
-        #   with 'buffer source array is read-only'.
-        kmeans_model = deepcopy(kmeans_model)
-        kmeans_model.model = self
-
-        # for i, cluster_model in enumerate(kmeans_model.cluster_models):
-        #     print(f"Model {i}: \t ", end=" ")
-        #     try:
-        #         print(cluster_model.cluster_centers_)
-        #     except AttributeError:
-        #         print("No cluster centers!")
-
-        # iter_coords = self.get_iter_coordinates(iteration)
-
-        kmeans_model.model.load_iter_data(iteration)
-        kmeans_model.model.get_transition_data_lag0()
-        # print(f"After loading coordPairList in iter {iteration}, shape is {kmeans_model.model.coordPairList.shape}")
-        parent_coords, child_coords = (
-            kmeans_model.model.coordPairList[..., 0],
-            self.coordPairList[..., 1],
-        )
-
-        # If there are no coords for this iteration, return None
-        if child_coords.shape[0] == 0:
-            return None, 0, iteration
-
-        # Otherwise, apply the k-means model and discretize
-        transformed_parent_coords = kmeans_model.model.coordinates.transform(
-            processCoordinates(parent_coords)
-        )
-        transformed_child_coords = kmeans_model.model.coordinates.transform(
-            processCoordinates(child_coords)
-        )
-
-        try:
-            kmeans_model.processing_from = True
-            try:
-                parent_dtrajs = kmeans_model.predict(transformed_parent_coords)
-            except IndexError as e:
-
-                print("Problem ===== ")
-                print(
-                    f"Parent pcoords are shape {kmeans_model.model.pcoord0List.shape}"
-                )
-                print(f"Parent coords are shape {transformed_parent_coords.shape}")
-                print(f"Child pcoords are shape {kmeans_model.model.pcoord1List.shape}")
-                print(f"Child coords are shape {transformed_child_coords.shape}")
-                print("===== ")
-
-                raise e
-
-            kmeans_model.processing_from = False
-            child_dtrajs = kmeans_model.predict(transformed_child_coords)
-        except AttributeError as e:
-            log.error("Cluster center was not initialized and not remapped")
-            log.error(kmeans_model.we_remap)
-            raise e
-            # TODO: Remap to nearest visited
-
-        return (
-            (parent_dtrajs, child_dtrajs),
-            1,
-            iteration,
-            kmeans_model.target_bins,
-            kmeans_model.basis_bins,
-        )
-
-    @ray.remote
-    def get_iter_fluxMatrix_ray(model, processCoordinates, n_iter):
-        """
-        TODO
-        ----
-        Refactor this so it doesn't need to take the ENTIRE model as an input. That serialization/deserialization is
-        painful and unnecessary.
-
-        Really, in the main loop, I can just pull out the data (pcoord lists + transition weights) for this iteration
-            and feed it directly in.
-        """
-
-        # model_id, n_iter, processCoordinates_id = args
-
-        # self = ray.get(model_id)
-        # processCoordinates = ray.get(processCoordinates_id)
-        self = model
-
-        # self.processCoordinates = processCoordinates
-
-        # self.clusters = deepcopy(self.clusters)
-        # self.clusters.model = self
-
-        iter_fluxmatrix = self.get_iter_fluxMatrix(n_iter)
-
-        del self
-
-        return iter_fluxmatrix, n_iter
-
-    def get_iter_fluxMatrix(self, n_iter):
-
-        self.load_iter_data(n_iter)
-        parent_pcoords = self.pcoord0List.copy()
-        child_pcoords = self.pcoord1List.copy()
-
-        self.get_transition_data_lag0()
-        transition_weights = self.transitionWeights.copy()
-
-        index_pairs = np.array(self.pair_dtrajs[n_iter - 1])
-
-        # Record every point where you're in the target
-        ind_end_in_target = np.where(self.is_WE_target(child_pcoords))
-
-        if ind_end_in_target[0].size > 0:
-            log.debug(
-                "Number of post-transition target1 entries: "
-                + str(ind_end_in_target[0].size)
-                + "\n"
-            )
-        else:
-            log.debug(f"No target1 entries. {ind_end_in_target}")
-
-        # Get the index of every point
-        ind_start_in_basis = np.where(self.is_WE_basis(parent_pcoords))
-        if ind_start_in_basis[0].size > 0:
-            log.debug(
-                "Number of pre-transition points in basis0: "
-                + str(ind_start_in_basis[0].size)
-                + "\n"
-            )
-
-        ind_end_in_basis = np.where(self.is_WE_basis(child_pcoords))
-        if ind_end_in_basis[0].size > 0:
-            log.debug(
-                "Number of post-transition points in basis1: "
-                + str(ind_end_in_basis[0].size)
-                + "\n"
-            )
-
-        return (
-            modelWE.build_flux_matrix(
-                self.n_clusters,
-                index_pairs,
-                ind_start_in_basis,
-                ind_end_in_basis,
-                ind_end_in_target,
-                transition_weights,
-            )
-            .todense()
-            .A
-        )
-
-    @ray.remote
-    def build_flux_matrix_remote(
-        n_clusters,
-        index_pairs,
-        ind_start_in_basis,
-        ind_end_in_basis,
-        ind_end_in_target,
-        transition_weights,
-        n_iter,
-    ):
-
-        return (
-            modelWE.build_flux_matrix(
-                n_clusters,
-                index_pairs,
-                ind_start_in_basis,
-                ind_end_in_basis,
-                ind_end_in_target,
-                transition_weights,
-            ),
-            n_iter,
-        )
-
-    # def get_iter_fluxMatrix(self, n_iter):
-    @staticmethod
-    def build_flux_matrix(
-        n_clusters,
-        index_pairs,
-        ind_start_in_basis,
-        ind_end_in_basis,
-        ind_end_in_target,
-        transition_weights,
-    ):
-        """
-        Build the flux matrix for an iteration.
-
-        Returns
-        -------
-
-        TODO
-        ----
-        This function is slow because of the call to get_transition_data_lag0(). See that function for more info.
-            Removed!
-
-        This can be refactored to be static
-        """
-
-        # 1. Update state with data from the iteration you want to compute the flux matrix for
-        # sys.stdout.write("iteration " + str(n_iter) + ": data \n")
-
-        # Need this to load pcoordXList
-        # self.load_iter_data(n_iter)
-        # parent_pcoords = self.pcoord0List.copy()
-        # child_pcoords = self.pcoord1List.copy()
-        #
-        # #  2. Load transition data at the requested lag
-        # if self.n_lag > 0:
-        #     raise NotImplementedError
-        #     # If you're using more than 1 lag, obtain segment histories at that lag
-        #     #   This means you have to look back one or more iterations
-        #
-        #     # sys.stdout.write("segment histories \n")
-        #     self.get_seg_histories(self.n_lag + 1)
-        #     # sys.stdout.write(" transition data...\n")
-        #     self.get_transition_data(self.n_lag)
-        # elif self.n_lag == 0:
-        #     # If you're using a lag of 0, your coordinate pairs are just the beginning/end of the iteration.
-        #     #   In that, it's not possible to have warps/recycles, since those are done handled between iterations.
-        #     try:
-        #         self.get_transition_data_lag0()
-        #         transition_weights = self.transitionWeights.copy()
-        #         pass
-        #     except KeyError:
-        #         log.warning(
-        #             f"No coordinates for iter {n_iter}, skipping this iter in fluxmatrix calculation"
-        #         )
-        #         return np.zeros(shape=(self.n_clusters + 2, self.n_clusters + 2))
-
-        # log.debug(f"Getting flux matrix for iter {n_iter}")# with {self.nSeg} segments")
-        # If you used a lag of 0, transitions weights are just weightList
-        # If you used a lag > 0, these include the weight histories from previous iterations
-        # num_transitions = np.shape(self.transitionWeights)[0]
-
-        # Create dedicated clusters for the target and basis states,
-        # and reassign any points within the target or basis to those
-        target_cluster_index = n_clusters + 1
-        basis_cluster_index = n_clusters
-
-        # (Segment, Atom, [lagged, current coord])
-        # log.debug(f"Coord pairlist shape is {self.coordPairList.shape}")
-
-        # Get which coords are not NaN
-        # good_coords = ~np.isnan(self.coordPairList).any(axis=(1, 2, 3))
-
-        # Assign a cluster to the lagged and the current coords
-
-        # reduced_initial, reduced_final = None, None
-        # reduced_initial = self.reduceCoordinates(
-        #     self.coordPairList[good_coords, :, :, 0]
-        # )
-
-        # reduced_final = self.reduceCoordinates(self.coordPairList[good_coords, :, :, 1])
-
-        # Wrap this in a try to make sure we unset the toggle
-        # Note: This toggle is only used for stratified clusteirng, but it's set either way.
-        # self.clusters.toggle = True
-        # self.clusters.processing_from = True
-        try:
-            # start_cluster = self.clusters.predict(reduced_initial)
-            # end_cluster = self.clusters.predict(reduced_final)
-            # start_cluster = np.array([x[0] for x in self.pair_dtrajs[n_iter-1]])
-            # end_cluster = np.array([x[1] for x in self.pair_dtrajs[n_iter-1]])
-            start_cluster, end_cluster = index_pairs.T.copy()
-        except Exception as e:
-            log.error(index_pairs)
-            # log.error(reduced_final)
-            # self.clusters.processing_from = False
-            # self.clusters.toggle = False
-            raise e
-        else:
-            pass
-            # self.clusters.processing_from = False
-            # self.clusters.toggle = False
-
-        # good_coords = np.arange(0, parent_pcoords.shape[0]).astype(int)
-
-        log.debug(f"Cluster 0 shape: {start_cluster.shape}")
-
-        log.debug(
-            f"Target cluster index is: {target_cluster_index},  basis cluster index is: {basis_cluster_index}"
-        )
-
-        # Re-assign points that were in either the target or the basis to the target or basis clusters
-        end_cluster[ind_end_in_target] = target_cluster_index
-        start_cluster[ind_start_in_basis] = basis_cluster_index
-        end_cluster[ind_end_in_basis] = basis_cluster_index
-
-        # coo_matrix takes in (Data, (x, y)) and then gives you a matrix, with the point at Data[i]
-        #   placed at (x,y)[i]
-        # Data here is just the number of segments since each segment is associated with 1 transition
-        try:
-            fluxMatrix = coo_matrix(
-                (transition_weights, (start_cluster, end_cluster)),
-                shape=(n_clusters + 2, n_clusters + 2),
-            )  # .todense()
-        except ValueError as e:
-            log.error(
-                f"Iter_fluxmatrix failed. Transition was from {start_cluster} -> {end_cluster} "
-                f"\n\t(Total {n_clusters + 2} clusters)"
-                f"\n\t(End in target: {ind_end_in_target})"
-                f"\n\t(Weights len: {len(transition_weights)})"
-                f"\n\t(Good Weights len: {len(transition_weights)})"
-                # f"\n\t(Good Coords len: {len(good_coords)})"
-                f"\n\t(start cluster len: {len(start_cluster)})"
-                f"\n\t(end_cluster len: {len(end_cluster)})"
-            )
-            raise e
-
-        # While the sparse matrix implementation is nice and efficient, using the np.matrix type is a little weird
-        #   and fragile, and it's possible it'll be deprecated in the future.
-        # fluxMatrix = fluxMatrix.A
-
-        return fluxMatrix
-
-    def get_fluxMatrix(
-        self,
-        n_lag,
-        first_iter=1,
-        last_iter=None,
-        iters_to_use=None,
-        use_ray=False,
-        result_batch_size=5,
-    ):
-        """
-        Compute the matrix of fluxes at a given lag time, for a range of iterations.
-
-        Checks if a file has been written named
-        "<`self.modelName`>_s<`first_iter`>_e<`last_iter`>_lag<`n_lag`>_clust<`self.n_clusters`>.h5".
-        If this file exists, load it and recalculate if it was calculated at an earlier iteration.
-        Otherwise, write it.
-
-        Updates:
-            - `self.n_lag`
-            - `self.errorWeight`
-            - `self.errorCount`
-            - `self.fluxMatrixRaw`
-
-        Parameters
-        ----------
-        n_lag: int
-            Number of lags to use.
-        first_iter: int
-            First iteration to use.
-        last_iter: int
-            Last iteration to use.
-
-        Returns
-        -------
-        None
-
-        """
-
-        self._fluxMatrixParams = [n_lag, first_iter, last_iter, iters_to_use]
-
-        if iters_to_use is not None:
-            log.debug(
-                "Specific iterations to use were provided for fluxmatrix calculation, using those."
-            )
-        elif iters_to_use is None:
-
-            if last_iter is None:
-                last_iter = self.maxIter
-            iters_to_use = range(first_iter + 1, last_iter)
-
-        # Else, if iters_to_use is Not none, and Not (first_iter is None and last_iter is None)
-        else:
-            log.error(
-                "Both specific iterations to use AND first/last were provided to fluxmatrix calculation. Specify"
-                " one or the other, not both."
-            )
-            raise Exception
-
-        log.debug("Computing all flux matrices")
-
-        self.n_lag = n_lag
-        self.errorWeight = 0.0
-        self.errorCount = 0
-
-        # +2 because the basis and target states are the last two indices
-        fluxMatrix = np.zeros((self.n_clusters + 2, self.n_clusters + 2))
-        nI = 0
-
-        fileName = (
-            self.modelName
-            + "-fluxmatrix-_s"
-            + str(first_iter)
-            + "_e"
-            + str(last_iter)
-            + "_lag"
-            + str(n_lag)
-            + "_clust"
-            + str(self.n_clusters)
-        )
-
-        # Overwrite this file, don't try to read from it.  Hence the "w" flag
-        # TODO: Maybe in the future return to this,
-        #  but it caused more problems than it was worth when doing multiple runs.
-        try:
-            f = h5py.File(fileName + ".h5", "w")
-        except BlockingIOError:
-            # Janky -- if the file exists, pick a random name
-            import random
-
-            fileName += f"_{int(random.random()*1000):04d}"
-            f = h5py.File(fileName + ".h5", "w")
-
-        dsetName = "fluxMatrix"
-
-        # FIXME: name this something descriptive or just use the 'in' statement in the if/elif
-        fluxmatrix_exists_in_h5 = dsetName in f
-        always_overwrite_fluxmatrix = True
-
-        # If this data file does not contain a fluxMatrix entry, create it
-        # For now, don't use a saved fluxmatrix, annoying to debug
-        if always_overwrite_fluxmatrix or not fluxmatrix_exists_in_h5:
-            # Create the fluxMatrix dataset
-            dsetP = f.create_dataset(dsetName, np.shape(fluxMatrix))
-            dsetP[:] = fluxMatrix
-
-            # Add up the flux matrices for each iteration to get the flux matrix.
-            # Then, save that matrix to the data file, along with the number of iterations used
-            # FIXME: Duplicated code
-            # The range is offset by 1 because you can't calculate fluxes for the 0th iteration
-            if not use_ray:
-                for iS in tqdm.tqdm(
-                    iters_to_use,
-                    desc="Constructing flux matrix",
-                ):
-                    log.debug("getting fluxMatrix iter: " + str(iS) + "\n")
-
-                    # fluxMatrixI = self.get_iter_fluxMatrix(iS)
-                    with concurrent.futures.ProcessPoolExecutor(
-                        max_workers=1, mp_context=mp.get_context("fork")
-                    ) as executor:
-                        fluxMatrixI = executor.submit(
-                            self.get_iter_fluxMatrix,
-                            iS,
-                        ).result()
-
-                    fluxMatrix = fluxMatrix + fluxMatrixI
-                    nI = nI + 1
-
-                    # f = h5py.File(fileName + ".h5", "a")
-                    dsetP = f[dsetName]
-                    dsetP[:] = fluxMatrix / nI
-                    dsetP.attrs["iter"] = iS
-                    log.debug(f"Completed flux matrix for iter {iS}")
-
-            # If we're running through Ray..
-            else:
-
-                # First, connect to the ray cluster
-                self.check_connect_ray()
-
-                # Submit all the tasks for iteration fluxmatrix calculations
-                task_ids = []
-
-                # model_id = ray.put(self)
-                # processCoordinates_id = ray.put(self.processCoordinates)
-
-                # max_inflight = 70
-                for iteration in tqdm.tqdm(
-                    iters_to_use,
-                    desc="Submitting fluxmatrix tasks",
-                ):
-
-                    # Allow 1000 in flight calls
-                    # This is buggy, so disable for now.
-                    # For example, if i = 5000, this call blocks until that
-                    # 4000 of the object_refs in result_refs are ready
-                    # and available.
-                    # See: https://docs.ray.io/en/latest/ray-design-patterns/limit-tasks.html
-                    # if len(task_ids) > max_inflight:
-                    #
-                    #     # The number that need to be ready before we can submit more
-                    #     num_ready = iteration - max_inflight
-                    #     log.info(f"At iteration {iteration}, waiting to submit more jobs until {num_ready} are ready")
-                    #     ready, notready = ray.wait(task_ids, num_returns=num_ready)
-                    #     log.info(f"At iteration {iteration}, {len(ready)} jobs are ready {len(notready)} not, submitting more.")
-
-                    # log.debug(f"Submitted fluxmatrix task iteration {iteration}")
-
-                    self.load_iter_data(iteration)
-                    parent_pcoords = self.pcoord0List.copy()
-                    child_pcoords = self.pcoord1List.copy()
-
-                    ind_end_in_target = np.where(self.is_WE_target(child_pcoords))
-                    ind_start_in_basis = np.where(self.is_WE_basis(parent_pcoords))
-                    ind_end_in_basis = np.where(self.is_WE_basis(child_pcoords))
-
-                    self.get_transition_data_lag0()
-                    transition_weights = self.transitionWeights.copy()
-
-                    index_pairs = np.array(self.pair_dtrajs[iteration - 1])
-
-                    # _id = self.get_iter_fluxMatrix_ray.remote(
-                    #     # model_id, processCoordinates_id, iteration
-                    #     parent_pcoords, child_pcoords, transition_weights
-                    # )
-
-                    _id = self.build_flux_matrix_remote.remote(
-                        self.n_clusters,
-                        index_pairs,
-                        ind_start_in_basis,
-                        ind_end_in_basis,
-                        ind_end_in_target,
-                        transition_weights,
-                        iteration,
-                    )
-
-                    task_ids.append(_id)
-
-                # Wait for them to complete
-                # Process results as they're ready, instead of in submission order
-                #  See: https://docs.ray.io/en/latest/ray-design-patterns/submission-order.html
-                # Additionally, this batches rather than getting them all at once, or one by one.
-
-                with tqdm.tqdm(
-                    total=len(iters_to_use), desc="Retrieving flux matrices"
-                ) as pbar:
-                    while task_ids:
-                        result_batch_size = min(result_batch_size, len(task_ids))
-                        log.debug(
-                            f"Waiting for {result_batch_size} results ({len(task_ids)} total remain)"
-                        )
-
-                        # Returns the first ObjectRefs that are ready, with a 60s timeout.
-                        finished, task_ids = ray.wait(
-                            task_ids, num_returns=result_batch_size, timeout=20
-                        )
-                        results = ray.get(finished)
-                        log.debug(f"Obtained {len(results)} results")
-
-                        # Add each matrix to the total fluxmatrix
-                        for _fmatrix, _iter in results:
-                            fluxMatrix = fluxMatrix + _fmatrix.todense().A
-                            pbar.update(1)
-                            pbar.refresh()
-
-                        # Try to free up some memory used by Ray for these objects
-                        # See: https://github.com/ray-project/ray/issues/15058
-                        # I was running into issues with objects spilling from the object store during fluxmatrix
-                        #   calculation. None of the individual calculations should really get that big, so maybe
-                        #   something wasn't getting freed from memory when it should've.
-                        del finished
-                        del results
-
-                log.info("Fluxmatrices all obtained")
-                # del model_id
-                del task_ids
-
-                # Write the H5. Can't do this per-iteration, because we're not guaranteed to be going sequentially now
-
-                nI = len(iters_to_use)
-                dsetP = f[dsetName]
-                dsetP[:] = fluxMatrix / nI
-                dsetP.attrs["iter"] = nI
-
-            f.close()
-
-            # Normalize the flux matrix by the number of iterations that it was calculated with
-            fluxMatrix = fluxMatrix / nI
-
-        # If this datafile DOES contain a fluxMatrix entry...
-        # HACK: This explicitly never runs now, but keep it around in case we decide to re-add this later.
-        # TODO: Update this to work with the specific iterations
-        elif fluxmatrix_exists_in_h5:
-
-            log.info("Fluxmatrix already exists in h5 file, loading saved.")
-
-            # Load the existing fluxMatrix
-            dsetP = f[dsetName]
-            fluxMatrix = dsetP[:]
-            nIter = dsetP.attrs["iter"]
-            # f.close()
-            nI = 1
-
-            # If the flux matrix was calculated at an earlier iteration than the current/requested last iteration, then
-            #   recalculate it and update it.
-            # TODO: Check if first_iter changed too? That's not stored anywhere
-
-            #  For now, just always overwrite any saved data.
-            # if nIter < last_iter:
-            if True:
-                log.warning(f"Ignoring any saved fluxmatrix in {fileName}.h5")
-                for iS in range(nIter, last_iter + 1):
-                    fluxMatrixI = self.get_iter_fluxMatrix(iS)
-                    fluxMatrix = fluxMatrix + fluxMatrixI
-                    nI = nI + 1
-                    # f = h5py.File(fileName + ".h5", "a")
-                    dsetP = f[dsetName]
-                    dsetP[:] = fluxMatrix / nI
-                    dsetP.attrs["iter"] = iS
-                    # f.close()
-                    sys.stdout.write("getting fluxMatrix iter: " + str(iS) + "\n")
-                f.close()
-                fluxMatrix = fluxMatrix / nI
-
-        f.close()
-
-        # Update state with the new, updated, or loaded from file fluxMatrix.
-        self.fluxMatrixRaw = fluxMatrix
-
-    def organize_fluxMatrix(self, use_ray=False, **args):
-        """
-        This cleaning step removes all clusters that aren't in the largest connected set, then rediscretizes all the
-        trajectories according to the new reduced set of clusters.
-        """
-
-        if not hasattr(self, "clustering_method"):
-            log.warning(
-                "self.clustering_method is not set. This may be a model saved before stratified was "
-                "implemented, or you may not have run cluster_coordinates! "
-                "Assuming the former and setting to aggregated."
-            )
-
-            self.clustering_method = "aggregated"
-
-        if self.clustering_method == "stratified":
-            self.organize_stratified(use_ray)
-
-            # TODO: Respect do_cleaning=False for blockwise stratified
-
-        elif self.clustering_method == "aggregated":
-
-            # If the user has manually specified states to keep, skip the normal cleaning, and just do what they said
-            if "states_to_keep" in args.keys():
-                self.organize_aggregated(use_ray=use_ray, **args)
-                return
-
-            # self.organize_aggregated(use_ray, **args)
-
-            fmatrix = self.fluxMatrixRaw.copy()
-            # Add recycling to avoid detecting the target as a sink
-            fmatrix[-1, -2] = 1.0
-
-            flat_raw_islands = np.sort(
-                [i for s in find_connected_sets(fmatrix, directed=True)[1:] for i in s]
-            )
-
-            new_args = {k: v for k, v in args.items()}
-            new_args["do_cleaning"] = False
-            states_to_keep = self.organize_aggregated(
-                use_ray=use_ray, **new_args
-            ).astype(bool)
-
-            regular_clean = np.argwhere(~states_to_keep)
-
-            if len(flat_raw_islands) > 0:
-                states_to_keep[flat_raw_islands] = False
-
-            states_to_keep[[self.n_clusters, self.n_clusters + 1]] = True
-
-            modified_clean = np.argwhere(~states_to_keep)
-
-            log.debug(
-                f"Modified cleaning added states {np.setdiff1d(modified_clean, regular_clean)}"
-            )
-
-            # If you didn't want to do cleaning, you've gone far enough, return the list of good state indices
-            if "do_cleaning" in args.keys() and not args["do_cleaning"]:
-                return np.argwhere(states_to_keep)
-
-            # Otherwise, carry on and actually clean
-            new_args = {k: v for k, v in args.items()}
-            new_args["states_to_keep"] = np.argwhere(states_to_keep)
-            self.organize_aggregated(use_ray=use_ray, **new_args)
-
-            return np.argwhere(states_to_keep)
-
-        else:
-            raise Exception(
-                f"Unrecognized clustering_method (Had: {self.clustering_method})"
-            )
-
-    def organize_aggregated(
-        self, use_ray=False, do_cleaning=True, states_to_keep=None, rediscretize=True
-    ):
-        """
-        Do some cleaning on the flux matrix, and update state with the cleaned flux matrix.
-
-        Namely:
-            - Remove unvisited clusters
-            - Remove bins with no connectivity
-            - Sort along the bins' projection in pcoord 1
-
-        Parameters
-        ----------
-        ray_args : dict
-            Dictionary whose values are currently unused, but containing an 'address' and 'password' key will enable
-            Ray.
-
-        do_cleaning : bool
-            If true, clean self.fluxMatrixRaw and re-discretize to produce self.fluxMatrix. If false, just return
-            the list of good states to keep.
-
-        Returns
-        -------
-        None
-
-        TODO
-        ----
-        Break this up into find_traps() and clean_traps(list_of_traps).
-        """
-
-        original_fluxmatrix = self.fluxMatrixRaw.copy()
-
-        log.debug("Cleaning flux matrix")
-
-        # Discretize trajectories via clusters
-        # Get the indices of the target and basis clusters
-        target_cluster_index = self.n_clusters + 1  # Target at -1
-        basis_cluster_index = self.n_clusters  # basis at -2
-
-        # This tracks which clusters are going to be cleaned from the flux matrix.
-        # A 0 means it'll be cleaned, a 1 means it'll be kept.
-        good_clusters = np.ones(self.n_clusters + 2)
-
-        cluster_pcoord_centers = np.zeros((self.n_clusters + 2, self.pcoord_ndim))
-        # cluster_pcoord_centers[indTargetCluster]=self.target_rmsd
-        cluster_pcoord_centers[target_cluster_index] = self.target_bin_center
-        # cluster_pcoord_centers[indBasisCluster]=self.get_reference_rmsd(self.basis_coords)
-        cluster_pcoord_centers[basis_cluster_index] = self.basis_bin_center
-
-        # Just initialize this to some positive nonzero value to kick off the while loop
-        nTraps = 1000
-        fluxMatrixTraps = self.fluxMatrixRaw.copy()
-        while nTraps > 0:
-            nTraps = 0
-            for cluster_index in range(self.n_clusters):
-                # Get the indices of the dtraj points in this cluster
-                idx_traj_in_cluster = [
-                    np.where(dtraj == cluster_index) for dtraj in self.dtrajs
-                ]
-
-                # Get the pcoord points that correspond to these dtraj points
-                offset = 0
-                pcoord_indices = []
-                for i, idxs in enumerate(idx_traj_in_cluster):
-                    pcoord_idxs = offset + idxs[0]
-                    pcoord_indices.extend(pcoord_idxs)
-                    offset += len(self.dtrajs[i])
-
-                # Get the number of dtraj points in this cluster
-                n_in_cluster = np.sum([x[0].shape for x in idx_traj_in_cluster])
-
-                # If this cluster is unvisited in the trajectories, set good_cluster to False
-                # if np.shape(idx_traj_in_cluster)[1] == 0:
-                if n_in_cluster == 0:
-                    good_clusters[cluster_index] = 0
-
-                # Otherwise, get the average pcoord of the points in the cluster
-                elif n_in_cluster > 0:
-                    # cluster_pcoord_centers[iC]=np.mean(self.get_reference_rmsd(self.coordSet[idx_traj_in_cluster[0],:,:]))
-                    # The coordinate of this cluster center is the average pcoord of all points in it
-
-                    try:
-                        cluster_pcoord_centers[cluster_index] = np.nanmean(
-                            self.pcoordSet[pcoord_indices, 0], axis=0
-                        )
-                    except IndexError as e:
-                        log.error(f"Offset is {offset}")
-                        log.error(pcoord_indices)
-                        raise e
-
-                # Get the total flux along the row and col of this index
-                net_flux = np.sum(fluxMatrixTraps[:, cluster_index]) + np.sum(
-                    fluxMatrixTraps[cluster_index, :]
-                )
-
-                # If both the row and column are all zero, this is an unvisited state, so set good_clusters to 0/False
-                if net_flux == 0.0:
-                    good_clusters[cluster_index] = 0
-
-                # If the sum along the row and column are nonzero
-                if net_flux > 0:
-                    # Get all the clusters that *aren't* the one we're looking at
-                    all_other_cluster_indices = np.setdiff1d(
-                        range(self.n_clusters), cluster_index
-                    )
-
-                    # Look at all the flux FROM other clusters
-                    total_flux_in = np.sum(
-                        fluxMatrixTraps[cluster_index, all_other_cluster_indices]
-                    )
-
-                    # And look at all the flux TO other clusters
-                    total_flux_out = np.sum(
-                        fluxMatrixTraps[all_other_cluster_indices, cluster_index]
-                    )
-
-                    # If either the flux from or the flux to other clusters are all zero,
-                    #   then this is a source or sink respectively.
-                    # So, clean it
-                    if total_flux_in == 0.0 or total_flux_out == 0.0:
-                        nTraps = nTraps + 1
-                        good_clusters[cluster_index] = 0
-                        fluxMatrixTraps[:, cluster_index] = 0.0
-                        fluxMatrixTraps[cluster_index, :] = 0.0
-
-            # Make sure we don't clean the target or basis clusters
-            good_clusters[basis_cluster_index] = 1
-            good_clusters[target_cluster_index] = 1
-
-        # If a list of states to keep has been provided, forget everything we just did, and only clean those.
-        # Still run the above, rather than putting the else for this if clause around it,
-        #   so that we still get the cluster pcoords from it.
-        if states_to_keep is not None:
-            good_clusters[:] = 0
-            good_clusters[states_to_keep] = 1
-
-        if not do_cleaning:
-            return good_clusters
-
-        # Store this array. 1 if a cluster is good, 0 otherwise.
-        clusters_good = np.copy(good_clusters)
-
-        # Get all the visited and non-trap clusters, which we want to keep, and make fluxMatrix from those
-        good_clusters = np.squeeze(np.where(good_clusters > 0))
-        fluxMatrix = self.fluxMatrixRaw[good_clusters, :]
-        fluxMatrix = fluxMatrix[:, good_clusters]
-
-        # Get the RMSD centers for all of the clusters we want to keep
-        cluster_pcoord_centers = cluster_pcoord_centers[good_clusters]
-        # Get the indices that sort it, in pcoord 0
-        pcoord_sort_indices = np.argsort(cluster_pcoord_centers[:, 0])
-        # And update the model's RMSD cluster centers to just include the sorted clusters to keep
-        self.targetRMSD_centers = cluster_pcoord_centers[pcoord_sort_indices]
-
-        # Sort fluxmatrix using the sorted indices, columns and then rows
-        fluxMatrix = fluxMatrix[pcoord_sort_indices, :]
-        fluxMatrix = fluxMatrix[:, pcoord_sort_indices]
-
-        # Renormalize the new flux matrix
-        self.fluxMatrix = fluxMatrix / np.sum(
-            fluxMatrix
-        )  # average weight transitioning or staying put should be 1
-
-        # Update the index of the basis and target states to account for their position in the new sorted clusters
-        originalClusters = good_clusters[pcoord_sort_indices]
-        # log.debug(f"Good clusters: {good_clusters}")
-        self.indBasis = np.where(originalClusters == basis_cluster_index)[0]
-        self.indTargets = np.where(originalClusters == target_cluster_index)[0]
-        log.debug(f"indBasis:  {self.indBasis}, indTargets: {self.indTargets}")
-        log.debug(
-            f"Sanity check -- basis:  {self.targetRMSD_centers[self.indBasis]}, target: {self.targetRMSD_centers[self.indTargets]}"
-        )
-
-        # Save the new, sorted clusters
-        self.originalClusters = originalClusters
-
-        # Update binCenters with the new, sorted centers
-        self.binCenters = cluster_pcoord_centers[pcoord_sort_indices]
-        self.nBins = np.shape(self.binCenters)[0]
-
-        # Remove the cluster structure dict entries corresponding to removed clusters
-        removed_clusters = np.argwhere(clusters_good == 0).squeeze()
-
-        # log.debug(f"Good clusters are {clusters_good}")
-        log.debug(f"Removed clusters were {removed_clusters}")
-        self.removed_clusters = removed_clusters
-
-        # TODO: Since I re-discretize below, this is no longer necessary.
-        #  For now, set it to 1:1 for backwards compatibility
-        # After cleaning, the cluster indices may no longer be consecutive.
-        #  So, when I do the cleaning, I need to build a mapping of old, nonconsecutive cluster indices from the
-        #  non-cleaned matrix, to new, consecutive indices.
-        # That's like collapsing the nonconsecutive list. In other words, imagine I started with 5 clusters [0,1,2,3,4]
-        #   and clean cluster 2. Now, I'll still have clusters labeled as [0,1,3,4], but my steady-state distribution
-        #   is only 4 elements. So indexing element 4 won't do anything.
-        cluster_mapping = {x: x for x in range(self.n_clusters + 2)}
-        n_removed = 0
-        for key in cluster_mapping.keys():
-            if key in self.removed_clusters:
-                n_removed += 1
-            cluster_mapping[key] = cluster_mapping[key] - n_removed
-
-        # TODO: This is unused, which is why it's just set to a 1:1. This is legacy, but left in for compatibility
-        #   with the restart plugin which expects it.
-        # log.debug(f"New cluster mapping is  {cluster_mapping}")
-        self.cluster_mapping = cluster_mapping
-        self.cluster_mapping = {
-            x: x for x in range(self.clusters.cluster_centers_.shape[0])
-        }
-
-        # Remove the bad clusters from the set of cluster centers. Thus, they won't be used by self.clusters.predict
-        #   in the future.
-        self.clusters.cluster_centers_ = np.delete(
-            self.clusters.cluster_centers_, removed_clusters, 0
-        )
-
-        # TODO: Don't duplicate code from cluster_coordinates!!
-        # TODO: You don't actually need to rediscretize every point -- just the removed ones.  Do this  later to make
-        #   this more efficient.
-        self.dtrajs = []
-
-        first_iter, last_iter = 1, self.maxIter
-
-        if rediscretize and not use_ray:
-            extra_iters_used = 0
-            for iteration in tqdm.tqdm(
-                range(first_iter, last_iter), desc="Post-cleaning rediscretization"
-            ):
-
-                if extra_iters_used > 0:
-                    extra_iters_used -= 1
-                    log.debug(f"Already processed  iter  {iteration}")
-                    continue
-
-                with concurrent.futures.ProcessPoolExecutor(
-                    max_workers=1, mp_context=mp.get_context("fork")
-                ) as executor:
-                    dtrajs, extra_iters_used = executor.submit(
-                        self.do_discretization,
-                        [self.clusters, iteration, self.processCoordinates],
-                    ).result()
-
-                    self.dtrajs.append(dtrajs)
-
-        # If we're using Ray...
-        elif rediscretize and use_ray:
-
-            # First, connect to the ray cluster
-            self.check_connect_ray()
-
-            # Submit all the discretization tasks to the cluster
-            task_ids = []
-
-            model_id = ray.put(self)
-            cluster_model_id = ray.put(self.clusters)
-            process_coordinates_id = ray.put(self.processCoordinates)
-
-            # max_inflight = 50
-            for iteration in tqdm.tqdm(
-                range(first_iter, last_iter), desc="Submitting discretization tasks"
-            ):
-                # if len(task_ids) > max_inflight:
-                #
-                #     # The number that need to be ready before we can submit more
-                #     num_ready = iteration - max_inflight
-                #     ray.wait(task_ids, num_returns=num_ready)
-
-                _id = self.do_ray_discretization.remote(
-                    model_id, cluster_model_id, iteration, process_coordinates_id
-                )
-                task_ids.append(_id)
-
-            # As they're completed, add them to dtrajs
-            dtrajs = [None] * (last_iter - first_iter)
-
-            # Do these in bigger batches, dtrajs aren't very big
-
-            with tqdm.tqdm(
-                total=len(task_ids), desc="Retrieving discretized trajectories"
-            ) as pbar:
-                while task_ids:
-                    result_batch_size = 50
-                    result_batch_size = min(result_batch_size, len(task_ids))
-
-                    # Returns the first ObjectRef that is ready, with a 20s timeout
-                    finished, task_ids = ray.wait(
-                        task_ids, num_returns=result_batch_size, timeout=20
-                    )
-                    results = ray.get(finished)
-
-                    for dtraj, _, iteration in results:
-                        dtrajs[iteration - 1] = dtraj
-                        pbar.update(1)
-                        pbar.refresh()
-
-            # Remove all empty elements from dtrajs and assign to self.dtrajs
-            self.dtrajs = [dtraj for dtraj in dtrajs if dtraj is not None]
-
-        self.removed_clusters = []
-
-        # Update self.n_clusters to account for any removed clusters
-        self.n_clusters -= n_removed
-
-        # Rebuild the fluxmatrix with whatever params were originally provided
-        self.get_fluxMatrix(*self._fluxMatrixParams, use_ray=use_ray)
-
-        # # The new, organized fluxmatrix is the result of computing the new fluxmatrix, on the new set of bins.
-        new_fluxMatrix = self.fluxMatrixRaw.copy()
-
-        # Do the same sorting and organization we did above on the "raw" fluxmatrix, for consistency.
-        new_fluxMatrix = new_fluxMatrix[pcoord_sort_indices, :]
-        new_fluxMatrix = new_fluxMatrix[:, pcoord_sort_indices]
-
-        # Renormalize the new flux matrix
-        new_fluxMatrix = new_fluxMatrix / np.sum(
-            new_fluxMatrix
-        )  # average weight transitioning or staying put should be 1
-
-        self.fluxMatrixRaw = original_fluxmatrix
-        self.fluxMatrix = new_fluxMatrix
-
-    def get_Tmatrix(self):
-        """
-        Compute the transition matrix from the flux matrix.
-        Corrects the "target" states to be true sink states.
-
-        More specifically:
-            - row-normalizes the flux matrix,
-            - sets any states with 0 flux ot (i.e. sinks) to have 1.0
-            - sets target bins to uniformly recycle into basis bins
-
-        Updates:
-            - `self.Tmatrix`
-
-        Returns
-        -------
-        None
-        """
-
-        log.debug("Computing transition matrix")
-
-        # Get a copy of the flux matrix
-        fluxmatrix = self.fluxMatrix.copy()
-        # Get the dimension of the flux matrix
-        fluxmatrix_shape = np.shape(fluxmatrix)
-        # Add up the total flux on each row, i.e. from each state
-        fluxes_out = np.sum(fluxmatrix, 1)
-
-        # For each state
-        for state_idx in range(fluxmatrix_shape[0]):
-            # For positive definite flux, set the matrix elements based on normalized fluxes
-            if fluxes_out[state_idx] > 0:
-                fluxmatrix[state_idx, :] = (
-                    fluxmatrix[state_idx, :] / fluxes_out[state_idx]
-                )
-
-            # If the flux is zero, then consider it all self-transition
-            # FIXME: this if can be an elif
-            if fluxes_out[state_idx] == 0.0:
-                fluxmatrix[state_idx, state_idx] = 1.0
-
-        # Make the transition matrix a steady-state matrix
-        # Identify the bins corresponding to target states.
-        sinkBins = self.indTargets  # np.where(avBinPnoColor==0.0)
-
-        # Get the number of sink bins
-        n_sink_bins = np.shape(sinkBins)
-        n_sink_bins = n_sink_bins[0]
-
-        # TODO: The goal here is to correct for the transition probabilities out of the sink state. Not
-        #   sure I fully understand this implementation, but that's the goal.
-        # Correction: I think this sets all the sink bins to recycle to the basis
-        sinkRates = np.zeros((1, self.nBins))
-        sinkRates[0, self.indBasis] = 1.0 / self.indBasis.size
-        tmatrix = fluxmatrix.copy()
-        tmatrix[sinkBins, :] = np.tile(sinkRates, (n_sink_bins, 1))
-
-        self.Tmatrix = tmatrix
-
-    def get_eqTmatrix(self):
-        Mt = self.fluxMatrix.copy()
-        n = np.shape(Mt)[0]
-        indSpace = np.arange(n).astype(int)
-        indSpace = np.setdiff1d(indSpace, np.append(self.indTargets, self.indBasis))
-        Mt = Mt[indSpace, :]
-        Mt = Mt[:, indSpace]
-        nR = np.shape(Mt)
-        sM = np.sum(Mt, 1)
-        for iR in range(nR[0]):
-            if sM[iR] > 0:
-                Mt[iR, :] = Mt[iR, :] / sM[iR]
-            if sM[iR] == 0.0:
-                Mt[iR, iR] = 1.0
-        self.Tmatrix = Mt
-
-    def get_steady_state(self, flux_fractional_convergence=1e-4, max_iters=10):
-        """ "
-        Get the steady-state distribution for the transition matrix.
-        Uses scipy eigensolver to obtain an initial guess, then refines that using inverse iteration.
-
-        Parameters
-        ----------
-        flux_fractional_convergence: (optional, default=1e-4) float
-            Convergence of the
-
-        max_iters: (optional, default=100) int
-
-        Notes
-        -----
-        Transition matrices generated from WE data may have entries spanning many orders of magnitude, and may have
-        extremely high condition numbers.
-        Furthermore, the smallest entries may be those near the target state, which are also the most important
-        for computing target fluxes, meaning values near machine precision can't just be truncated.
-        All this means that floating-point error may substantially affect the results of eigensolvers, and may produce
-        bad/negative/inaccurate values for small probability bins.
-
-        In order to obtain better estimates, sparse matrices are used to reduce the number of floating point operations
-        being performed.
-        A stationary distribution is first estimated using scipy's eigensolver for sparse matrices.
-        This is then used as an initial guess for the inverse iteration method, to further refine it.
-        Convergence of the inverse iteration is determined using change in the flux estimate.
-        """
-
-        # Cast the matrix to a sparse matrix, to reduce floating point operations
-        sparse_mat = sparse.csr_matrix(self.Tmatrix)
-
-        algebraic_pss = self.get_steady_state_algebraic(
-            max_iters=10, check_negative=False, set=False
-        )
-
-        # Get an initial flux estimate using the numpy algebraic solver
-        # Call with _set=False so you don't actually update self.JtargetSS
-        last_flux = self.get_steady_state_target_flux(pSS=algebraic_pss, _set=False)
-
-        # ## Next, use that as an initial guess  for inverse iteration
-        last_pSS = algebraic_pss
-
-        log.debug(f"Initial flux: {last_flux}\n")
-        flux_warned = False
-
-        for N in range(max_iters):
-
-            iterated = inverse_iteration(matrix=sparse_mat, guess=last_pSS)
-
-            # Compute change in pSS
-            pSS_change = np.sqrt(np.mean(np.power(iterated - last_pSS, 2)))
-            log.debug(f"\t Change in SS: {pSS_change:.2e}")
-            last_pSS = iterated
-
-            # Compute change in target flux with the new pSS
-            new_flux = self.get_steady_state_target_flux(pSS=last_pSS, _set=False)
-
-            flux_change = new_flux - last_flux
-            log.debug(
-                f"\t Change in flux estimate: {flux_change:.2e} \t ({new_flux:.2e} raw)"
-            )
-            last_flux = new_flux
-
-            # Set the convergence criterion after the iteration, in case the original result is really junk
-            # Do it after so the first time this is calculated, it's using the result of the first iteration.
-            # If the initial numpy estimate was really bad, it'll never be used.
-            flux_convergence_criterion = last_flux * flux_fractional_convergence
-            log.debug(f"\t Flux convergence criterion is {flux_convergence_criterion}")
-
-            if N > 0:
-                if last_flux == 0 and not flux_warned:
-                    log.warning(
-                        "Flux is 0, so steady-state solver will only converge after max iterations (using flux as "
-                        "convergence criterion is not meaningful if it's 0!). "
-                        "If you're looking for equilibrium, this is probably OK."
-                        " Otherwise, take a look at why you have 0 flux."
-                    )
-                    flux_warned = True
-
-            if abs(flux_change) < flux_convergence_criterion:
-                log.info(
-                    f"\nFlux converged to {last_flux:.4e} after {N + 1} iterations of inverse iteration."
-                )
-                break
-
-            elif N == max_iters - 1 and not last_flux == 0:
-                log.warning("Flux is nonzero and did not converge!")
-
-        assert (last_pSS >= 0).all(), "Negative elements in pSS"
-        assert last_flux >= 0, "Negative flux estimate from this pSS"
-
-        log.info("Done with steady-state estimation.")
-        self.pSS = last_pSS
-
-    def get_steady_state_algebraic(self, max_iters=1000, check_negative=True, set=True):
-        """
-        Compute the steady-state distribution as the eigenvectors of the transition matrix.
-
-        Updates:
-            - `self.pSS`
-
-        Parameters
-        ----------
-        max_iters: int, optional
-            Defaults to 1000. Number of power-method iterations to run if the numpy eigensolver returns negative elements.
-
-        check_negative: bool, optional
-            Defaults to True. If True, then raise an Exception if there are negative elements in the normalized pSS.
-
-        Returns
-        -------
-        None
-        """
-
-        # log.warning(
-        #     "get_steady_state_algebraic() will be deprecated soon. Use get_steady_state() instead, which has"
-        #     " a more robust eigensolver."
-        # )
-
-        log.debug("Computing steady-state from eigenvectors")
-
-        eigenvalues, eigenvectors = np.linalg.eig(np.transpose(self.Tmatrix))
-
-        pSS = np.real(eigenvectors[:, np.argmax(np.real(eigenvalues))])
-
-        pSS = pSS.squeeze()
-
-        assert not np.isclose(np.sum(pSS), 0), "Steady-state distribution sums to 0!"
-        pSS = pSS / np.sum(pSS)
-
-        # The numpy eigensolver is iterative, and approximate. Given that data from WE often spans many orders of
-        #   magnitude, we'll sometimes run into situations where our populations span more than machine precision.
-        #   This causes hiccups in the eigensolver. However, we can't just zero these out (as attempted above),
-        #   because these values are often important.
-        # So, if there are any negative elements, try to correct the NP eigensolver result using the matrix method
-        if sum(pSS < 0) > 0 and max_iters > 0:
-            log.info(
-                "Negative elements in pSS after normalization, attempting to correct with matrix power method."
-            )
-            pSS_last = pSS
-            _tmatrix = self.Tmatrix.copy()
-
-            for N in range(max_iters):
-
-                pSS_new = _tmatrix.T @ pSS_last
-                num_negative_elements = sum(pSS_new < 0)
-                if num_negative_elements == 0:
-                    log.info(f"Corrected to semidefinite pSS in {N} iterations")
-                    break
-
-                pSS_last = pSS_new
-                _tmatrix = np.matmul(self.Tmatrix, _tmatrix)
-
-            if N == max_iters - 1:
-                log.warning(
-                    "Power method did NOT obtain semidefinite pSS. Some negative values remain. This is weird, and you"
-                    " should try to figure out why this is happening."
-                )
-            else:
-                pSS = pSS_new
-
-        elif max_iters == 0:
-            log.debug(
-                "Negative elements in pSS after normalization, "
-                "but max_iters=0 so not attempting to correct with power method."
-            )
-
-        if not np.all(pSS >= 0):
-
-            if check_negative:
-                assert np.all(
-                    pSS >= 0
-                ), f"Some negative elements in steady-state distribution: {pSS}"
-            else:
-                log.warning(
-                    "Some negative  elements in pSS... Ignoring, and setting model.pSS anyways."
-                )
-
-        if set:
-            self.pSS = pSS
-        else:
-            return pSS
-
-    def get_steady_state_matrixpowers(self, conv):
-        """
-        Compute the steady-state distribution using the matrix power method.
-
-        Updates:
-            - `self.pSS`
-
-        Parameters
-        ----------
-        conv: numeric
-            Convergence criterion for iteration.
-
-        Returns
-        -------
-        None
-        """
-        max_iters = 10000
-        Mt = self.Tmatrix.copy()
-        dconv = 1.0e100
-        N = 1
-        pSS = np.mean(Mt, 0)
-        pSSp = np.ones_like(pSS)
-        while dconv > conv and N < max_iters:
-            Mt = np.matmul(self.Tmatrix, Mt)
-            N = N + 1
-            if N % 10 == 0:
-                pSS = np.mean(Mt, 0)
-                pSS = pSS / np.sum(pSS)
-                dconv = np.sum(np.abs(pSS - pSSp))
-                pSSp = pSS.copy()
-                sys.stdout.write("N=" + str(N) + " dconv: " + str(dconv) + "\n")
-                self.pSS = pSS.copy()
-
-    def get_steady_state_target_flux(self, pSS=None, _set=True):
-        """
-        Get the total flux into the target state(s).
-
-        Updates:
-            - `self.lagtime`
-            - `self.JtargetSS`
-
-        Parameters
-        ----------
-        pSS: (optional) array-like
-            Steady-state distribution. If nothing provided, then use self.pSS
-
-        _set: (optional) boolean
-            If True, then update self.JtargetSS and self.lagtime. If False, then just return the value of JtargetSS.
-
-        Returns
-        -------
-        None
-        """
-
-        # If this matrix isn't connected from source to target, then it's not possible to get an MFPT directly out of it
-        sparse_mat = sparse.csr_matrix(self.Tmatrix)
-        connected = is_connected(
-            sparse_mat, self.indBasis, self.indTargets, directed=True
-        )
-        if not connected:
-            log.critical(
-                "There is no path in this matrix from the basis to the target, so no MFPT can be calculated."
-            )
-            return -1
-
-        Mss = self.Tmatrix
-
-        # If no pSS was provided, then pull from self
-        if pSS is None:
-            pSS = np.squeeze(np.array(self.pSS))
-
-        lagtime = self.tau * (self.n_lag + 1)
-
-        # Get a list of all the states that AREN'T targets, since we want to sum up
-        nTargets = self.indTargets.size
-        indNotTargets = np.setdiff1d(range(self.nBins), self.indTargets)
-        # log.debug(f"Non-target states are those with index {indNotTargets}")
-
-        Jt = 0.0
-        # Add up the total flux into each of the targets
-        for j in range(nTargets):
-            log.debug(
-                f"Processing flux into target state with index {self.indTargets[j]}"
-            )
-
-            jj = self.indTargets[j]
-
-            Jt = Jt + np.sum(
-                np.multiply(
-                    pSS[indNotTargets],
-                    np.squeeze(
-                        np.array(Mss[indNotTargets, jj * np.ones_like(indNotTargets)])
-                    ),
-                )
-            )
-
-        if _set:
-            self.lagtime = lagtime
-            self.JtargetSS = Jt / self.lagtime
-        else:
-            return Jt / lagtime
-
-    def get_flux(self):
-        """
-        Get the measured flux (i.e. from the flux matrix) into the target.
-        """
-
-        # J = np.zeros_like(self.binCenters)
-        nBins = np.shape(self.targetRMSD_centers)[0]
-        J = np.zeros(nBins)
-        fluxMatrix = self.fluxMatrix.copy()
-        for i in range(0, nBins - 1):
-            indBack = range(i + 1)
-            indForward = range(i + 1, nBins)
-            JR = 0.0
-            JF = 0.0
-            for j in indBack:
-                JR = JR + np.sum(fluxMatrix[indForward, j * np.ones_like(indForward)])
-            for j in indForward:
-                JF = JF + np.sum(fluxMatrix[indBack, j * np.ones_like(indBack)])
-            J[i] = JR - JF
-            self.J = J
-
-    def get_flux_committor(self):
-        """
-        Get the flux binned according to committors
-
-        Returns
-        -------
-
-        """
-
-        # J = np.zeros_like(self.targetRMSD_centers)
-        # nBins = np.shape(self.binCenters)[0]
-        nBins = np.shape(self.targetRMSD_centers)[0]
-        J = np.zeros(nBins)
-        fluxMatrix = self.fluxMatrix.copy()
-        indq = np.argsort(np.squeeze(1.0 - self.q))
-        fluxMatrix = fluxMatrix[indq, :]
-        fluxMatrix = fluxMatrix[:, indq]
-
-        for i in tqdm.tqdm(
-            range(0, nBins - 1), desc="Obtaining pseudocommittor-fluxes"
-        ):
-            indBack = range(i + 1)
-            indForward = range(i + 1, nBins)
-            JR = 0.0
-            JF = 0.0
-            for j in indBack:
-                JR = JR + np.sum(fluxMatrix[indForward, j * np.ones_like(indForward)])
-            for j in indForward:
-                JF = JF + np.sum(fluxMatrix[indBack, j * np.ones_like(indBack)])
-            J[i] = JR - JF
-            self.Jq = J.squeeze() / self.tau
-            # sys.stdout.write("%s " % i)
-
-    def plot_flux_committor_pcoordcolor(
-        self,
-        nwin=1,
-        ax=None,
-        pcoord_to_use=0,
-        **_plot_args,
-    ):
-
-        _models = [self]
-        _model_labels = ["main_model"]
-
-        plot_args = {
-            "linewidth": 2,
-            "s": 50,
-            "marker": ">",
-            "cmap": plt.cm.rainbow.reversed(),
-            "alpha": 0.7,
-        }
-
-        plot_args.update(_plot_args)
-
-        if ax is None:
-            fig = plt.figure(figsize=(10, 4))
-            ax = fig.add_subplot(111)
-
-        for i, (_model, _label) in enumerate(zip(_models[::-1], _model_labels[::-1])):
-
-            if _model is None:
-                continue
-
-            if not hasattr(_model, "q"):
-                log.warning(
-                    f"Committors have not yet been generated for {_label}, generating now."
-                )
-                _model.get_committor()
-
-            if not hasattr(_model, "Jq"):
-                log.warning(
-                    f"Committor-fluxes have not yet been generated for {_label}, generating now."
-                )
-                _model.get_flux_committor()
-
-            n_bins = _model.targetRMSD_centers.shape[0]
-            Jq_avg = _model.Jq.copy()
-            Jq_std = np.zeros_like(Jq_avg)
-
-            q_avg = np.zeros_like(Jq_avg)
-
-            indq = np.argsort(np.squeeze(1.0 - _model.q))
-            for _i in range(n_bins - 1, nwin - 1, -1):
-                iav = _i - nwin
-                ind = range(_i - nwin, _i)
-                Jq_avg[iav] = np.mean(_model.Jq[ind])
-                Jq_std[iav] = np.std(_model.Jq[ind])
-                q_avg[iav] = np.mean(_model.q[indq[ind]])
-
-            indPlus = np.where(Jq_avg > 0.0)
-
-            lines = ax.scatter(
-                q_avg[indPlus],
-                np.squeeze(Jq_avg[indPlus]),
-                c=_model.targetRMSD_centers[indPlus, pcoord_to_use],
-                label=f"{_label} flux toward target",
-                **plot_args,
-            )
-
-        print("Plotting committor")
-        ax.figure.colorbar(lines, label=f"Progress Coordinate {pcoord_to_use}")
-
-        ax.set_xlim([-0.1, 1.1])
-
-        ax.set_title("Full-data model")
-        ax.set_yscale("log")
-        ax.set_xlabel("Pseudocommittor")
-        ax.set_ylabel("Flux (weight/second)")
-        self.print_pseudocommittor_warning()
-
-        return ax, lines
-
-    def plot_flux_committor(
-        self,
-        nwin=1,
-        ax=None,
-        save=False,
-        suppress_validation=False,
-        _from_colors=None,
-        _to_colors=None,
-        **_plot_args,
-    ):
-
-        _models = [self]
-        _models.extend(self.validation_models)
-
-        _model_labels = ["main_model"]
-        _model_labels.extend(
-            [f"validation_model_{n}" for n in range(len(self.validation_models))]
-        )
-
-        cross_validation_done = (
-            self.validation_models is not None and len(self.validation_models) > 1
-        )
-        if not cross_validation_done:
-            log.critical(
-                "No cross-validation models have been generated! Do this before making plots."
-            )
-            if not suppress_validation:
-                raise Exception("Perform cross-validation before plotting results.")
-
-        own_ax = False
-        if ax is None:
-            own_ax = True
-            fig = plt.figure(figsize=(10, 4))
-            ax = fig.add_subplot(111)
-
-        plot_args = {"linewidth": 2, "markersize": 10}
-        plot_args.update(_plot_args)
-
-        n_models = len(_models)
-        from_colors = [
-            plt.get_cmap("cool")(0.25 + (0.75 * i / n_models)) for i in range(n_models)
-        ]
-        to_colors = [
-            plt.get_cmap("hot")(0.25 + (0.5 * i / n_models)) for i in range(n_models)
-        ]
-
-        for i, (_model, _label) in enumerate(zip(_models, _model_labels)):
-
-            if _model is None:
-                continue
-
-            if not hasattr(_model, "q"):
-                log.warning(
-                    f"Committors have not yet been generated for {_label}, generating now."
-                )
-                _model.get_committor()
-
-            if not hasattr(_model, "Jq"):
-                log.warning(
-                    f"Committor-fluxes have not yet been generated for {_label}, generating now."
-                )
-                _model.get_flux_committor()
-
-            n_bins = _model.targetRMSD_centers.shape[0]
-            Jq_avg = _model.Jq.copy()
-            Jq_std = np.zeros_like(Jq_avg)
-
-            q_avg = np.zeros_like(Jq_avg)
-
-            indq = np.argsort(np.squeeze(1.0 - _model.q))
-            for _i in range(n_bins - 1, nwin - 1, -1):
-                iav = _i - nwin
-                ind = range(_i - nwin, _i)
-                Jq_avg[iav] = np.mean(_model.Jq[ind])
-                Jq_std[iav] = np.std(_model.Jq[ind])
-                q_avg[iav] = np.mean(_model.q[indq[ind]])
-
-            indPlus = np.where(Jq_avg > 0.0)
-            indMinus = np.where(Jq_avg < 0.0)
-
-            if _from_colors is not None:
-                plot_args["color"] = _from_colors[i]
-            else:
-                plot_args["color"] = from_colors[i]
-
-            ax.plot(
-                q_avg[indMinus],
-                -np.squeeze(Jq_avg[indMinus]),
-                "<",
-                # color=to_target_colors[i%len(to_target_colors)],
-                label=f"{_label} flux toward source/basis",
-                **plot_args,
-            )
-
-            if _to_colors is not None:
-                plot_args["color"] = _to_colors[i]
-            else:
-                plot_args["color"] = to_colors[i]
-
-            ax.plot(
-                q_avg[indPlus],
-                np.squeeze(Jq_avg[indPlus]),
-                ">",
-                # color=from_target_colors[i%len(from_target_colors)],
-                label=f"{_label} flux toward target",
-                **plot_args,
-            )
-
-        ax.set_yscale("log")
-        ax.set_xscale("linear")
-        ax.set_xlim([-0.1, 1.1])
-        ax.set_xlabel("Pseudocommittor")
-        ax.set_ylabel("Flux (weight/second)")
-        self.print_pseudocommittor_warning()
-
-        if own_ax:
-            ax.legend(bbox_to_anchor=(1.01, 1.0), loc="upper left")
-            fig.tight_layout()
-
-        if save:
-            plot_filename = f"{self.modelName}_flux_committor.pdf"
-            log.info(f"Saving flux-committor plot to {plot_filename}")
-            plt.savefig(plot_filename)
-
-        return ax
-
-    def plot_flux(
-        self,
-        custom_name=None,
-        ax=None,
-        save=False,
-        suppress_validation=False,
-        _from_colors=None,
-        _to_colors=None,
-        pcoord_to_use=0,
-        **_plot_args,
-    ):
-        """
-        Make, and save, a plot of the fluxes along the RMSD.  get_flux() must be run before this.
-
-        Parameters
-        ----------
-        custom_name : str (optional)
-            The name for the saved plot. Defaults to flux_s<first iter>_e<last iter>.png
-
-        Returns
-        -------
-
-        """
-
-        _models = [self]
-        _models.extend(self.validation_models)
-
-        _model_labels = ["main_model"]
-        _model_labels.extend(
-            [f"validation_model_{n}" for n in range(len(self.validation_models))]
-        )
-
-        cross_validation_done = (
-            self.validation_models is not None and len(self.validation_models) > 1
-        )
-        if not cross_validation_done:
-            log.critical(
-                "No cross-validation models have been generated! Do this before making plots."
-            )
-            if not suppress_validation:
-                raise Exception("Perform cross-validation before plotting results.")
-
-        own_ax = False
-        if ax is None:
-            own_ax = True
-            fig = plt.figure(figsize=(10, 4))
-            ax = fig.add_subplot(111)
-
-        plot_args = {"linewidth": 2, "markersize": 10}
-        plot_args.update(_plot_args)
-
-        n_models = len(_models)
-        from_colors = [
-            plt.get_cmap("cool")(0.25 + (0.75 * i / n_models)) for i in range(n_models)
-        ]
-        to_colors = [
-            plt.get_cmap("hot")(0.25 + (0.5 * i / n_models)) for i in range(n_models)
-        ]
-
-        # Draw the basis/target boundaries in this pcoord
-        [
-            ax.axvline(
-                bound, color="r", linestyle="--", label=["", "Target boundary"][i == 0]
-            )
-            for i, bound in enumerate(self.target_pcoord_bounds[pcoord_to_use, :])
-        ]
-        [
-            ax.axvline(
-                bound,
-                color="b",
-                linestyle="--",
-                label=["", "Basis/Source boundary"][i == 0],
-            )
-            for i, bound in enumerate(self.basis_pcoord_bounds[pcoord_to_use, :])
-        ]
-
-        for i, (_model, _label) in enumerate(zip(_models, _model_labels)):
-
-            if _model is None:
-                continue
-
-            if not hasattr(_model, "J"):
-                log.warning(
-                    f"Fluxes have not yet been generated for {_label}, generating now."
-                )
-                _model.get_flux()
-
-            J = _model.J / _model.tau
-
-            binCenters = _model.targetRMSD_centers[:, pcoord_to_use]
-            indPlus = np.where(J > 0.0)
-            indMinus = np.where(J < 0.0)
-
-            if _to_colors is not None:
-                plot_args["color"] = _to_colors[i]
-            else:
-                plot_args["color"] = to_colors[i]
-
-            ax.plot(
-                binCenters[indPlus],
-                np.squeeze(J[indPlus]),
-                ">",
-                label=f"{_label} flux toward target",
-                **plot_args,
-            )
-
-            if _from_colors is not None:
-                plot_args["color"] = _from_colors[i]
-            else:
-                plot_args["color"] = from_colors[i]
-
-            ax.plot(
-                binCenters[indMinus],
-                -np.squeeze(J[indMinus]),
-                "<",
-                label=f"{_label} flux toward source/basis",
-                **plot_args,
-            )
-
-        ax.set_yscale("log")
-        ax.set_xlabel(f"Pcoord {pcoord_to_use}")
-        ax.set_ylabel("Flux (weight/second)")
-
-        if own_ax:
-            ax.legend(bbox_to_anchor=(1.01, 1.0), loc="upper left")
-            fig.tight_layout()
-
-        if save:
-            if custom_name is not None:
-                plot_filename = custom_name
-            else:
-                plot_filename = f"{self.modelName}_flux.pdf"
-            log.info(f"Saving flux plot to {plot_filename}")
-            plt.savefig(plot_filename)
-
-        return ax
-
-    def evolve_target_flux(self):
-        Mss = self.Tmatrix
-        probTransient = self.probTransient
-        nT = np.shape(probTransient)[0]
-        Jtarget = np.zeros(nT)
-        self.lagtime = self.tau * (self.n_lag + 1)
-        nTargets = self.indTargets.size
-        indNotTargets = np.setdiff1d(range(self.nBins), self.indTargets)
-        JtargetTimes = np.zeros(nT)
-        for iT in range(nT):
-            Jt = 0.0
-            for j in range(nTargets):
-                jj = self.indTargets[j]
-                Jt = Jt + np.sum(
-                    np.multiply(
-                        probTransient[iT, indNotTargets],
-                        Mss[indNotTargets, jj * np.ones_like(indNotTargets)],
-                    )
-                )
-            Jtarget[iT] = Jt
-            JtargetTimes[iT] = iT * self.nStore * self.lagtime
-        self.Jtarget = Jtarget / self.lagtime
-        self.JtargetTimes = JtargetTimes
-
-    def get_committor(self, conv=1e-5):
-        """
-        Iteratively obtain an estimate of the committor.
-
-        1. Take the flux matrix, and normalize it into a transition matrix.
-
-        2. Apply two-sided absorbing boundary conditions by setting self-transition probabilities for the basis and
-            target states to 1.0, and all transitions out to 0.0.
-
-        3. Starting with an initial committor "guess" of all 1s, iteratively multiply the guess by the transition matrix
-            until convergence is below conv.
-
-        Updates:
-            - self.q
-
-        Parameters
-        ----------
-        conv: numerical
-            Convergence criteria for committor calculation. Calculation stops when the total difference between q_p and q
-            is less than this.
-
-        Returns
-        -------
-
-        """
-
-        self.print_pseudocommittor_warning()
-
-        _fluxMatrix = self.fluxMatrix.copy()
-
-        # Number of bins/states in the fluxmatrix
-
-        num_bins = np.shape(_fluxMatrix)
-        flux_out = np.sum(_fluxMatrix, 1)
-
-        for bin_idx in range(num_bins[0]):
-
-            # If the flux out is positive semidefinite, then normalize that row to get transition probabilities
-            if flux_out[bin_idx] > 0:
-                _fluxMatrix[bin_idx, :] = _fluxMatrix[bin_idx, :] / flux_out[bin_idx]
-
-            # If the flux out is zero, then just set the self-transition to 1.0
-            # (This probably keeps a calculation later clean?)
-            if flux_out[bin_idx] == 0.0:
-                _fluxMatrix[bin_idx, bin_idx] = 1.0
-
-        #  The basis states are set to have 0 flux out, and only self-transition = 1.0.
-        #   These are the two-sided absorbing BCs for the committor.
-        sinkBins = self.indBasis
-        for ii in sinkBins:
-            _fluxMatrix[ii, :] = np.zeros((1, self.nBins))
-            _fluxMatrix[ii, ii] = 1.0
-
-        q = np.zeros((self.nBins, 1))
-
-        # Committor to the target state is 1 by definition
-        q[self.indTargets, 0] = 1.0
-
-        dconv = 100.0
-        progress = 0
-        qp = np.ones_like(q)
-
-        # Iteratively update the committor estimate until it converges to stationarity
-        # (The committor is the stationary distribution for two-sided absorbing boundary conditions)
-        with tqdm.tqdm(total=-np.log10(conv)) as pbar:
-            while dconv > conv:
-                q[self.indTargets, 0] = 1.0
-                q[self.indBasis, 0] = 0.0
-                q = np.matmul(_fluxMatrix, q)
-                dconv = np.sum(np.abs(qp - q))
-
-                # Update with progress since last iter
-                log.debug("convergence: " + str(dconv) + "\n")
-                pbar.update(max(0, -np.log10(dconv)) - progress)
-                progress = max(0, -np.log10(dconv))
-
-                qp = q.copy()
-                self.q = q
-
-        self.q = q.squeeze()
-
-    # TODO: This should probably just be a call to get_committor, followed by self.q = 1 - self.q
-    def get_backwards_committor(self, conv):
-        Mt = self.fluxMatrix.copy()
-        nR = np.shape(Mt)
-        sM = np.sum(Mt, 1)
-        for iR in range(nR[0]):
-            if sM[iR] > 0:
-                Mt[iR, :] = Mt[iR, :] / sM[iR]
-            if sM[iR] == 0.0:
-                Mt[iR, iR] = 1.0
-        sinkBins = self.indTargets  # np.where(avBinPnoColor==0.0)
-        nsB = np.shape(sinkBins)
-        nsB = nsB[0]
-        for ii in sinkBins:
-            Mt[ii, :] = np.zeros((1, self.nBins))
-            Mt[ii, ii] = 1.0
-        Mt = np.transpose(Mt)  # time reversal
-        q = np.zeros((self.nBins, 1))
-        q[self.indBasis, 0] = 1.0
-        dconv = 100.0
-        qp = np.ones_like(q)
-        while dconv > conv:
-            q[self.indBasis, 0] = 1.0
-            q[self.indTargets, 0] = 0.0
-            q = np.matmul(Mt, q)
-            dconv = np.sum(np.abs(qp - q))
-            sys.stdout.write("convergence: " + str(dconv) + "\n")
-            qp = q.copy()
-            self.qm = q
-        self.q = q.copy()
-
-    def plot_committor(self):
-        fig = plt.figure(figsize=(8, 6))
-        plt.scatter(self.targetRMSD_centers[:, 0], self.q, s=15, c="black")
-        plt.yscale("log")
-        plt.ylabel("Pseudocommittor to target", fontsize=12)
-        plt.xlabel("Average microstate pcoord", fontsize=12)
-
-        self.print_pseudocommittor_warning()
-
-        plt.pause(1)
-        fig.savefig(
-            self.modelName
-            + "_s"
-            + str(self.first_iter)
-            + "_e"
-            + str(self.last_iter)
-            + "committor.png"
-        )
 
     @staticmethod
     def print_pseudocommittor_warning():
